% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-transforms.R
\name{acid_coord_flip}
\alias{acid_coord_flip}
\title{Flip coordinates}
\usage{
acid_coord_flip(object)
}
\arguments{
\item{object}{\code{ggplot}.}
}
\value{
\code{ggplot}.
}
\description{
Intelligently flip a plot with discrete data on the X axis.
}
\details{
This function puts the samples that were near the left origin on the X at the
top on the Y axis, making them more human readable.
}
\note{
Updated 2019-07-29.
}
\examples{
library(ggplot2)
g <- ggplot(data = mpg, aes(x = class)) + geom_bar()
print(g)
acid_coord_flip(g)
}
\seealso{
\itemize{
\item \code{coord_flip()}.
}
}
