% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-layers.R
\name{acid_geom_abline}
\alias{acid_geom_abline}
\title{Reference line}
\usage{
acid_geom_abline(xintercept = NULL, yintercept = NULL)
}
\arguments{
\item{xintercept, yintercept}{\code{numeric(1)}.
Value denoting x- or y-axis cutoff.
Use either \code{xintercept} or \code{yintercept}, but not both.}
}
\value{
\code{Layer}/\code{ggproto}.
}
\description{
Add a horizontal (using \code{yintercept}) or vertical (using \code{xintercept})
reference line.
}
\note{
Updated 2023-08-10.
}
\examples{
## x-axis line.
geom <- acid_geom_abline(xintercept = 1L)
print(geom)

## y-axis line.
geom <- acid_geom_abline(yintercept = 1L)
print(geom)
}
