% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-layers.R
\name{acid_geom_bar}
\alias{acid_geom_bar}
\title{Bar chart}
\usage{
acid_geom_bar(..., color = NA, stat = "identity")
}
\arguments{
\item{...}{Additional arguments, passed to \code{ggplot2::geom_bar()}.}

\item{color}{\code{character(1)}.
Line color.
Defaults to disabled, using \code{NA}.}

\item{stat}{\code{character(1)}.
Statistical transformation to use on the data for this layer.}
}
\value{
\code{Layer}/\code{ggproto}.
}
\description{
Modified version of \code{ggplot2::geom_bar()}.
}
\note{
Updated 2021-06-29.
}
\examples{
data <- data.frame(
    x = c("a", "b", "c", "d"),
    y = c(5L, 10L, 15L, 20L)
)
# Note that `.data` is a data pronoun from the rlang package.
p <- ggplot(data = data, mapping = aes(x = .data[["x"]], y = .data[["y"]]))
p + acid_geom_bar(fill = "black")
}
