% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-layers.R
\name{acid_geom_label}
\alias{acid_geom_label}
\title{Text label}
\usage{
acid_geom_label(data = NULL, mapping = NULL, ...)
}
\arguments{
\item{data}{\code{data.frame}.
Data frame, containing plot data.}

\item{mapping}{Set of aesthetic mappings created by \code{ggplot2::aes()}.}

\item{...}{Additional arguments, passed to \code{ggplot2::geom_label()}.}
}
\value{
\code{Layer}/\code{ggproto}.
}
\description{
Modified version of \code{ggplot2::geom_label()}.
}
\note{
Updated 2021-06-29.
}
\examples{
geom <- acid_geom_label()
print(geom)
}
