% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-layers.R
\name{acid_geom_label_average}
\alias{acid_geom_label_average}
\title{Average labels}
\usage{
acid_geom_label_average(data, col, fun = c("mean", "median"), digits = 0L, ...)
}
\arguments{
\item{col}{\code{character(1)}.
Column name.}

\item{fun}{\code{character(1)}.
Function name to use for average calculation.
Currently supports mean or median.}

\item{digits}{\code{integer(1)}.
Number of significant digits to use.
Defaults to rounded.}

\item{...}{Additional arguments, passed to \code{\link[=acid_geom_label]{acid_geom_label()}}.}
}
\value{
\code{Layer}/\code{ggproto}.
}
\description{
Add average labels to a plot.
}
\details{
For example, \code{col} can be \code{nGene}. Median or mean values are always
calculated per sample (\code{sampleName}).
}
\note{
Updated 2021-06-29.
}
\examples{
data = data.frame(
    "sampleName" = rep(c("sample1", "sample2"), times = 4L),
    "counts" = seq_len(8L)
)
geom <- acid_geom_label_average(
    data = data,
    col = "counts",
    fun = "mean"
)
print(geom)
}
