% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-layers.R
\name{acid_geom_label_repel}
\alias{acid_geom_label_repel}
\title{Repulsive textual annotations}
\usage{
acid_geom_label_repel(
  data = NULL,
  mapping = NULL,
  color = NULL,
  size = 4L,
  ...
)
}
\arguments{
\item{data}{\code{data.frame}.
Data frame, containing plot data.}

\item{mapping}{Set of aesthetic mappings created by \code{ggplot2::aes()}.}

\item{color}{\code{character(1)} or \code{NULL}.
Text color.}

\item{size}{\code{integer(1)}.
Font size.}

\item{...}{Additional arguments, passed to \code{ggrepel::geom_label_repel()}.}
}
\value{
\code{Layer}/\code{ggproto}.
}
\description{
Modified version of \code{ggrepel::geom_label_repel()}.
}
\details{
If advanced customization of the text labels is required, simply use the
ggrepel version instead.
}
\note{
Updated 2021-06-29.
}
\examples{
geom <- acid_geom_label_repel()
print(geom)
}
