% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-breaks.R
\name{acid_pretty_breaks}
\alias{acid_pretty_breaks}
\title{Pretty breaks that resemble the base plotting engine}
\usage{
acid_pretty_breaks(n = 5L)
}
\arguments{
\item{n}{\code{integer(1)}.
Desired number of breaks.}
}
\value{
\code{function}.
}
\description{
Pretty breaks that resemble the base plotting engine
}
\note{
Updated 2019-07-29.
}
\examples{
acid_pretty_breaks()
}
\seealso{
\itemize{
\item \code{\link[grDevices:axisTicks]{grDevices::axisTicks()}}.
\item \code{\link[scales:breaks_log]{scales::log_breaks()}}.
\item \code{\link[scales:trans_breaks]{scales::trans_breaks()}}.
\item Adapted from a Stack Overflow post:
https://stackoverflow.com/questions/14255533
}
}
