% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-transforms.R
\name{acid_scale_y_continuous_nopad}
\alias{acid_scale_y_continuous_nopad}
\title{Remove Y axis padding}
\usage{
acid_scale_y_continuous_nopad(..., expand = expand_scale(mult = c(0L,
  0.025)))
}
\arguments{
\item{...}{Other arguments passed on to \code{scale_(x|y)_continuous()}}

\item{expand}{Vector of range expansion constants used to add some
padding around the data, to ensure that they are placed some distance
away from the axes. Use the convenience function \code{\link[=expand_scale]{expand_scale()}}
to generate the values for the \code{expand} argument. The defaults are to
expand the scale by 5\% on each side for continuous variables, and by
0.6 units on each side for discrete variables.}
}
\value{
\code{ggproto}.
}
\description{
Remove Y axis padding
}
\note{
Updated 2019-07-29.
}
\examples{
acid_scale_y_continuous_nopad()
}
