% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{acidplots-package}
\alias{acidplots}
\alias{acidplots-package}
\title{acidplots
Functions for plotting genomic data.}
\description{
Functions for plotting genomic data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://acidplots.acidgenomics.com/}
  \item \url{https://github.com/acidgenomics/acidplots/}
  \item Report bugs at \url{https://github.com/acidgenomics/acidplots/issues/}
}

}
\author{
\strong{Maintainer}: Michael Steinbaugh \email{mike@steinbaugh.com} (\href{https://orcid.org/0000-0002-2403-2221}{ORCID})

Other contributors:
\itemize{
  \item Acid Genomics [copyright holder, funder]
}

}
\keyword{internal}
