% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-scales.R
\name{autoScale}
\alias{autoScale}
\alias{acid_scale_color_continuous}
\alias{acid_scale_fill_continuous}
\alias{acid_scale_colour_discrete}
\alias{acid_scale_fill_discrete}
\alias{autoContinuousColorScale}
\alias{autoContinuousFillScale}
\alias{autoDiscreteColorScale}
\alias{autoDiscreteFillScale}
\title{Automatic color/fill scales to use with ggplot object}
\usage{
acid_scale_color_continuous()

acid_scale_fill_continuous()

acid_scale_colour_discrete()

acid_scale_fill_discrete()
}
\value{
\code{Scale}/\code{ggproto}/\code{gg}.
}
\description{
Helper functions for setting default colors in Acid Genomics plots.
}
\note{
Updated 2023-08-11.
}
\section{Setting default colors}{


These functions will inherit values defined in global \code{options()}. Either
input a ggplot2 color scale function that returns \code{Scale} class
(e.g. \code{scale_color_synesthesia_c}), or can input \code{"gradient"}, \code{"viridis"}
character strings.
\itemize{
\item \code{acid_scale_color_continuous()}: \code{"ggplot2.continuous.colour"}.
\item \code{acid_scale_color_discrete()}: \code{"ggplot2.discrete.colour"}.
\item \code{acid_scale_fill_continuous()}: \code{"ggplot2.continuous.fill"}.
\item \code{acid_scale_fill_discrete()}: \code{"ggplot2.discrete.fill"}.
}
}

\examples{
x <- acid_scale_color_continuous()
class(x)
x <- acid_scale_color_discrete()
class(x)
x <- acid_scale_fill_continuous()
class(x)
x <- acid_scale_fill_discrete()
class(x)
}
\seealso{
\itemize{
\item https://ggplot2.tidyverse.org/reference/scale_colour_continuous.html
\item https://ggplot2.tidyverse.org/reference/scale_colour_discrete.html
}
}
