% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/barcodeRanksPerSample-methods.R
\name{barcodeRanksPerSample}
\alias{barcodeRanksPerSample}
\alias{barcodeRanksPerSample,SingleCellExperiment-method}
\title{Barcode ranks per sample}
\usage{
barcodeRanksPerSample(object, ...)

\S4method{barcodeRanksPerSample}{SingleCellExperiment}(object, assay = 1L, ...)
}
\arguments{
\item{object}{Object.}

\item{assay}{\code{vector(1)}.
Assay name or index position.}

\item{...}{Passthrough arguments to \code{DropletUtils::barcodeRanks()}.}
}
\value{
\code{DataFrameList}.
}
\description{
Barcode ranks per sample
}
\note{
Requires DropletUtils package to be installed.

Updated 2021-09-13.
}
\examples{
data(SingleCellExperiment, package = "AcidTest")

## SingleCellExperiment ====
if (requireNamespace("DropletUtils", quietly = TRUE)) {
    object <- SingleCellExperiment
    x <- barcodeRanksPerSample(object)
    names(x)
}
}
\seealso{
\itemize{
\item \code{DropletUtils::barcodeRanks()}.
}
}
