% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breaks.R
\name{base_breaks}
\alias{base_breaks}
\title{Pretty breaks that resemble the base plotting engine}
\usage{
base_breaks(n = 5L)
}
\arguments{
\item{n}{\code{integer(1)}.
Desired number of breaks.}
}
\value{
\code{function}.
}
\description{
Pretty breaks that resemble the base plotting engine
}
\examples{
base_breaks()
}
\seealso{
\itemize{
\item \code{\link[grDevices:axisTicks]{grDevices::axisTicks()}}.
\item \code{\link[scales:log_breaks]{scales::log_breaks()}}.
\item \code{\link[scales:trans_breaks]{scales::trans_breaks()}}.
\item Adapted from a Stack Overflow post:
https://stackoverflow.com/questions/14255533
}
}
