% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-transforms.R
\name{flip_x_discrete}
\alias{flip_x_discrete}
\title{Flip discrete data}
\usage{
flip_x_discrete(object)
}
\arguments{
\item{object}{\code{ggplot}.}
}
\value{
\code{ggplot}.
}
\description{
Intelligently flip a plot with discrete data on the X axis.
}
\details{
This function puts the samples that were near the left origin on the X at the
top on the Y axis, making them more human readable.
}
\examples{
library(ggplot2)
g <- ggplot(data = mpg, aes(x = class)) + geom_bar()
print(g)
flip_x_discrete(g)
}
\seealso{
\itemize{
\item \code{coord_flip()}.
}
}
