% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genderMarkers}
\alias{genderMarkers}
\title{Sexually dimorphic gender marker genes}
\format{An object of class \code{list} of length 2.}
\usage{
genderMarkers
}
\description{
Sexually dimorphic gender marker genes
}
\examples{
data(genderMarkers)
genderMarkers
}
\keyword{datasets}
