% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-layers.R
\name{geoms}
\alias{geoms}
\alias{acid_geom_abline}
\alias{acid_geom_bar}
\alias{acid_geom_label}
\alias{acid_geom_label_average}
\alias{acid_geom_label_repel}
\title{ggplot2 geometric objects}
\usage{
acid_geom_abline(xintercept = NULL, yintercept = NULL)

acid_geom_bar(..., color = NA, stat = "identity")

acid_geom_label(data = NULL, mapping = NULL, ...)

acid_geom_label_average(data, col, fun = c("mean", "median"), digits = 0L, ...)

acid_geom_label_repel(
  data = NULL,
  mapping = NULL,
  color = NULL,
  size = 4L,
  ...
)
}
\arguments{
\item{xintercept, yintercept}{\code{numeric(1)}.
Value denoting x- or y-axis cutoff. Specify one but not both.}

\item{...}{Additional arguments.}

\item{color}{\code{character(1)}.
Text color (e.g. orange).}

\item{stat}{\code{character(1)}.
Statistical transformation to use on the data for this layer.}

\item{data}{\code{data.frame}.
Data.}

\item{mapping}{Set of aesthetic mappings created by \code{ggplot2::aes()}.}

\item{col}{\code{character(1)}.
Column name.}

\item{fun}{\code{character(1)}.
Function name to use for average calculation.
Currently supports mean or median.}

\item{digits}{\code{integer(1)}.
Number of significant digits to use.
Defaults to rounded.}

\item{size}{\code{integer(1)}.
Font size.}
}
\value{
\code{ggproto}.
}
\description{
Convenience functions with modified defaults for
\href{http://ggplot2.org}{ggplot2}.
}
\note{
Updated 2021-06-29.
}
\section{acid_geom_abline}{


Horizontal or vertical cutoff line.
}

\section{acid_geom_label}{


Modified version of \code{ggplot2::geom_label()}.
}

\section{acid_geom_label_average}{


Add average labels to a plot. For example, \code{col} can be \code{nGene}. Median or
mean values are always calculated per sample (\code{sampleName}).
}

\section{acid_geom_label_repel}{


Repulsive textual annotations. Modified basejump version of
\code{ggrepel::geom_label_repel()}. If advanced customization of the text labels
is required, simply use the ggrepel version instead.
}

\examples{
## acid_geom_abline ====
## x-axis line
geom <- acid_geom_abline(xintercept = 1L)
geom

## y-axis line
geom <- acid_geom_abline(yintercept = 1L)
geom

## acid_geom_label ====
geom <- acid_geom_label()
geom

## acid_geom_label_average ====
data = data.frame(
    sampleName = rep(c("sample1", "sample2"), times = 4L),
    counts = seq_len(8L)
)
geom <- acid_geom_label_average(
    data = data,
    col = "counts",
    fun = "mean"
)
geom

## acid_geom_label_repel ====
geom <- acid_geom_label_repel()
geom
}
\seealso{
\itemize{
\item \code{ggplot2::geom_label()}.
\item \code{ggrepel::geom_label_repel()}.
}
}
