% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient.R
\name{gradient}
\alias{gradient}
\title{Hex color gradient generator}
\usage{
gradient(colors = c(low = "darkorchid3", mid = "gray50", high =
  "darkorange2"), n = 256L)
}
\arguments{
\item{colors}{\code{character}.
Color names or hexadecimal values used to define the gradient.}

\item{n}{\code{integer(1)}.
The number of colors (>= 1) to be in the palette.}
}
\value{
\code{character}.
Hexadecimal colors in RGB space.
}
\description{
Hex color gradient generator
}
\examples{
gradient(n = 3L)
}
