% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{isDark}
\alias{isDark}
\title{Should the plot labels be rendered in dark mode?}
\usage{
isDark()
}
\value{
\code{logical(1)}.
}
\description{
Should the plot labels be rendered in dark mode?
}
\note{
Updated 2019-09-13.
}
\examples{
isDark()
}
