% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchLabels.R
\name{matchLabels}
\alias{matchLabels}
\title{Match ggplot2 labels}
\usage{
matchLabels(labels, choices)
}
\arguments{
\item{labels}{\code{list}.
User-defined plot labels.}

\item{choices}{\code{list}.
Default plot labels, defined in \code{labels} \code{\link[base:formals]{formals()}}.}
}
\description{
Allow user to define ggplot labels, and populate missing values from the
defaults specified in \code{labels} formal argument.
}
\note{
Attempting to follow a similar naming convention to
\code{\link[base:match.arg]{match.arg()}} here, using \code{choices} argument.

Updated 2019-09-16.
}
\examples{
matchLabels(
    labels = list(title = "XXX"),
    choices = list(
        title = "AAA",
        subtitle = "BBB"
    )
)
}
\seealso{
\code{\link[ggplot2:labs]{ggplot2::labs()}}.
}
