% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{minimalism-package}
\alias{minimalism}
\alias{minimalism-package}
\title{minimalism
Minimalist visualizations for bioinformatics.}
\description{
Part of the basejump toolkit for bioinformatics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://minimalism.acidgenomics.com/}
  \item \url{https://github.com/acidgenomics/minimalism/}
  \item Report bugs at \url{https://github.com/acidgenomics/minimalism/issues/}
}

}
\author{
\strong{Maintainer}: Michael Steinbaugh \email{mike@steinbaugh.com} (0000-0002-2403-2221)

Other contributors:
\itemize{
  \item Acid Genomics [copyright holder, funder]
}

}
\keyword{internal}
