% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\docType{data}
\name{minimalismTestsURL}
\alias{minimalismTestsURL}
\title{freerange test data URL}
\format{An object of class \code{character} of length 1.}
\usage{
minimalismTestsURL
}
\description{
freerange test data URL
}
\examples{
minimalismTestsURL
}
\keyword{internal}
