% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{Parameters}
\arguments{
\item{aspectRatio}{\code{integer(1)}.
Aspect ratio.}

\item{assay}{\code{vector(1)}.
Name or index of count matrix slotted in
\code{\link[SummarizedExperiment:assays]{assays()}}. When passing in a string, the
name must be defined in
\code{\link[SummarizedExperiment:assayNames]{assayNames()}}.}

\item{color}{\code{ScaleDiscrete}.
Desired ggplot2 color scale. Must supply discrete values. When set \code{NULL},
the default ggplot2 color palette will be used. If manual color definitions
are desired, we recommend using \code{\link[ggplot2:scale_color_manual]{ggplot2::scale_color_manual()}}.

To set the discrete color palette globally, use:\preformatted{options(acid.color.discrete = ggplot2::scale_color_viridis_d())
}}

\item{countsAxisLabel}{\code{character(1)}.
Counts axis label.}

\item{dark}{\code{logical(1)}.
Plot against a dark background using the
\code{\link[acidplots:theme_midnight]{theme_midnight()}} ggplot2 theme.}

\item{dimsUse}{\code{integer}.
Vector of length 2 that denotes the columns from the reduced dimension
matrix to use for \code{centerX} and \code{centerY} column calculations. Defaults the
first and second dimensions.}

\item{direction}{\code{character(1)}.
Plot "\code{both}", "\code{up}", or "\code{down}" directions.}

\item{expression}{\code{character(1)}.
Calculation to apply.
Uses \code{\link[base:match.arg]{match.arg()}} internally and defaults to the first
argument in the \code{character} vector.}

\item{fill}{\code{ggproto}/\code{ScaleDiscrete}.
Desired ggplot2 fill scale. Must supply discrete values. When set to
\code{NULL}, the default ggplot2 color palette will be used. If manual color
definitions are desired, we recommend using \code{\link[ggplot2:scale_fill_manual]{ggplot2::scale_fill_manual()}}.

To set the discrete fill palette globally, use:\preformatted{options(acid.fill.discrete = ggplot2::scale_fill_viridis_d())
}}

\item{flip}{\code{logical(1)}.
Flip x and y axes. Recommended for plots containing many samples.}

\item{gene2symbol}{\code{Gene2Symbol}.
Gene-to-symbol mappings. Must contain \code{geneID} and \code{geneName} columns. See
\code{Gene2Symbol} for more information.}

\item{genes}{\code{character}.
Gene identifiers. It is considered better practice to input the stable gene
identifiers from Ensembl (e.g. "ENSG00000000003") and not the (HGNC) gene
symbols (e.g. "TSPN6"), if possible.}

\item{geom}{\code{character(1)}.
Plot type. Uses \code{\link[base:match.arg]{match.arg()}} internally and defaults to
the first argument in the \code{character} vector.}

\item{headerLevel}{\code{integer(1)} (\code{1}-\code{7}).
Markdown header level.}

\item{interestingGroups}{\code{character}.
Groups of interest that define the samples. If left unset, defaults to
\code{sampleName}.}

\item{label}{\code{logical(1)}.
Superimpose sample text labels on the plot.}

\item{labelSize}{\code{integer(1)}.
Size of the text label.}

\item{legend}{\code{logical(1)}.
Show plot legend.}

\item{level}{\code{character(1)}.
Return as genes or transcripts.}

\item{limit}{\code{numeric(1)}.
Threshold to denote on the plot, using a dashed line.}

\item{max}{\code{numeric(1)}.
Recommended maximum value cutoff.}

\item{min}{\code{numeric(1)}.
Recommended minimum value cutoff.}

\item{minCounts}{\code{integer(1)}.
Minimum number of counts per gene in the count matrix.}

\item{n}{\code{integer(1)}.
Number to include.}

\item{ntop}{\code{integer(1)}.
Number of top genes to label.}

\item{object}{Object.}

\item{perMillion}{\code{logical(1)}.
Display as counts per million.}

\item{perSample}{\code{logical(1)}.
Visualize the distributions per sample.}

\item{plotlist}{\code{list}.
List containing \code{ggplot} objects.}

\item{pointAlpha}{\code{numeric(1)} (\code{0}-\code{1}).
Alpha transparency level. Useful when there many points in the dataset
(e.g. single-cell data), and some points can be masked.}

\item{pointColor}{\code{character(1)}.
Default point color for the plot.}

\item{pointScalar}{\code{integer(1)}.
Default point size for the plot.}

\item{pointsAsNumbers}{\code{logical(1)}.
Plot the points as numbers (\code{TRUE}) or dots (\code{FALSE}).}

\item{pointSize}{\code{numeric(1)}.
Point size for dots in the plot.}

\item{return}{\code{character(1)}.
Return type. Uses \code{\link[base:match.arg]{match.arg()}} internally and defaults
to the first argument in the \code{character} vector.}

\item{sigPointColor}{\code{character}.
Color names for labeling upregulated and downregulated genes. Also supports
a character string for labeling DEGs with the same color, regardless of
direction.}

\item{subtitle}{\code{character(1)}.
Plot subtitle.}

\item{title}{\code{character(1)}.
Plot title.}

\item{trans}{\code{character(1)}.
Name of the axis scale transformation to apply.

For more information:\preformatted{help(topic = "scale_x_continuous", package = "ggplot2")
}}

\item{trendline}{\code{logical(1)}.
Include trendline on plot.}

\item{...}{Additional arguments.}
}
\value{
No value.
}
\description{
Parameters
}
\keyword{internal}
