% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBarcodeRanks-methods.R
\docType{methods}
\name{plotBarcodeRanks}
\alias{plotBarcodeRanks}
\alias{plotBarcodeRanks,SingleCellExperiment-method}
\title{Plot barcode ranks}
\usage{
plotBarcodeRanks(object, ...)

\S4method{plotBarcodeRanks}{SingleCellExperiment}(object,
  colors = c(fitline = acidplots::lightPalette[["blue"]], inflection =
  acidplots::lightPalette[["purple"]], knee =
  acidplots::lightPalette[["orange"]]), ...)
}
\arguments{
\item{object}{Object.}

\item{colors}{\code{character(3)}.
Character vector denoting \code{fitline}, \code{inflection}, and \code{knee} point colors.
Must pass in color names or hexadecimal values.}

\item{...}{Passthrough to \code{\link[=barcodeRanksPerSample]{barcodeRanksPerSample()}}.}
}
\value{
\code{ggplot} grid.
}
\description{
Plot barcode ranks
}
\note{
Updated 2019-09-15.
}
\examples{
data(SingleCellExperiment, package = "acidtest")

## SingleCellExperiment ====
## Not supported for R < 3.6.
if (requireNamespace("DropletUtils", quietly = TRUE)) {
    object <- SingleCellExperiment
    plotBarcodeRanks(object)
}
}
