% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotCellCounts-methods.R
\name{plotCellCounts}
\alias{plotCellCounts}
\alias{plotCellCounts,SingleCellExperiment-method}
\title{Plot cell counts}
\usage{
plotCellCounts(object, ...)

\S4method{plotCellCounts}{SingleCellExperiment}(
  object,
  interestingGroups = NULL,
  labels = list(title = "Cell counts", subtitle = NULL, x = NULL, y = "cells")
)
}
\arguments{
\item{object}{Object.}

\item{interestingGroups}{\code{character}.
Groups of interest to use for visualization.
Corresponds to factors describing the columns of the object.}

\item{labels}{\code{list}.
ggplot2 labels.
See \code{\link[ggplot2:labs]{ggplot2::labs()}} for details.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Plot cell counts
}
\note{
Updated 2022-03-05.
}
\examples{
data(SingleCellExperiment_splatter, package = "AcidTest")

## SingleCellExperiment ====
object <- SingleCellExperiment_splatter
plotCellCounts(object)
}
\author{
Michael Steinbaugh, Rory Kirchner
}
