% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/plotCellCountsPerCluster-methods.R
\name{plotCellCountsPerCluster}
\alias{plotCellCountsPerCluster}
\alias{plotCellCountsPerCluster,SingleCellExperiment-method}
\title{Plot cell counts per cluster}
\usage{
plotCellCountsPerCluster(object, ...)

\S4method{plotCellCountsPerCluster}{SingleCellExperiment}(object, interestingGroups = NULL)
}
\arguments{
\item{object}{Object.}

\item{interestingGroups}{\code{character}.
Groups of interest to use for visualization.
Corresponds to factors describing the columns of the object.}

\item{...}{Additional arguments.}
}
\value{
Show graphical output. Invisibly return \code{ggplot}.
}
\description{
Plot cell counts per cluster
}
\note{
Updated 2023-08-10.
}
\examples{
data(SingleCellExperiment_Seurat, package = "AcidTest")

## SingleCellExperiment ====
object <- SingleCellExperiment_Seurat
plotCellCountsPerCluster(object)
}
