% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/plotCellTypesPerCluster-methods.R
\name{plotCellTypesPerCluster}
\alias{plotCellTypesPerCluster}
\alias{plotCellTypesPerCluster,SingleCellExperiment,KnownMarkers-method}
\title{Plot cell types per cluster}
\usage{
plotCellTypesPerCluster(object, markers, ...)

\S4method{plotCellTypesPerCluster}{SingleCellExperiment,KnownMarkers}(
  object,
  markers,
  min = 1L,
  max = Inf,
  reduction = "UMAP",
  expression = c("mean", "sum"),
  headerLevel = 2L,
  ...
)
}
\arguments{
\item{object}{Object.}

\item{markers}{Object containing gene markers.}

\item{min}{\code{numeric(1)}.
Recommended minimum value cutoff.}

\item{max}{\code{numeric(1)}.
Recommended maximum value cutoff.}

\item{reduction}{\code{vector(1)}.
Dimension reduction name or index position.}

\item{expression}{\code{character(1)}.
Calculation to apply.
Uses \code{\link[base:match.arg]{match.arg()}} internally and defaults to the first
argument in the \code{character} vector.}

\item{headerLevel}{\code{integer(1)} (\code{1}-\code{7}).
Markdown header level.}

\item{...}{Passthrough arguments to \code{\link[=plotMarker]{plotMarker()}}.}
}
\value{
Show graphical output. Invisibly return \code{list}.
}
\description{
Plot cell types per cluster
}
\details{
Plot the geometric mean of the significant marker genes for every known cell
type (per unbiased cluster). Cell types with too few (\code{min} cutoff) or too
many (\code{max} cutoff) marker genes will be skipped.
}
\note{
Updated 2023-08-17.
}
\examples{
data(
    KnownMarkers,
    SingleCellExperiment_Seurat,
    package = "AcidTest"
)

## SingleCellExperiment ====
object <- SingleCellExperiment_Seurat
markers <- KnownMarkers
plotCellTypesPerCluster(
    object = object,
    markers = markers,
    reduction = "UMAP"
)
}
