% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotCleveland-methods.R
\name{plotCleveland}
\alias{plotCleveland}
\alias{plotCleveland,DFrame-method}
\alias{plotCleveland,data.frame-method}
\title{Correlation X-Y scatterplot}
\usage{
plotCleveland(object, ...)

\S4method{plotCleveland}{DFrame}(
  object,
  xCol,
  yCol,
  labels = list(title = NULL, subtitle = NULL, x = NULL, y = NULL),
  color = "black"
)

\S4method{plotCleveland}{data.frame}(
  object,
  xCol,
  yCol,
  labels = list(title = NULL, subtitle = NULL, x = NULL, y = NULL),
  color = "black"
)
}
\arguments{
\item{object}{Object.}

\item{xCol, yCol}{\code{character(1)} or \code{integer(1)}.
X and Y column name or position.}
}
\value{
Plot.
}
\description{
Correlation X-Y scatterplot
}
\note{
Updated 2024-04-05.
}
\examples{
## FIXME Need to add working example.
}
