% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotCounts-methods.R
\name{plotCounts}
\alias{plotCounts}
\alias{plotDots}
\alias{plotViolin}
\alias{plotCounts,SingleCellExperiment-method}
\alias{plotCounts,SummarizedExperiment-method}
\alias{plotDots,SingleCellExperiment-method}
\alias{plotViolin,SingleCellExperiment-method}
\title{Plot counts}
\usage{
plotCounts(object, ...)

plotDots(object, ...)

plotViolin(object, ...)

\S4method{plotCounts}{SingleCellExperiment}(
  object,
  genes,
  assay = c("logcounts", "normcounts"),
  geom = c("violin", "dot"),
  perSample = TRUE,
  legend = getOption(x = "acid.legend", default = TRUE),
  title = NULL
)

\S4method{plotCounts}{SummarizedExperiment}(
  object,
  genes,
  assay = 1L,
  interestingGroups = NULL,
  convertGenesToSymbols = TRUE,
  geom = c("point", "violin", "boxplot", "bar"),
  trans = c("identity", "log2", "log10"),
  line = c("none", "median", "mean", "geometricMean"),
  legend = getOption(x = "acid.legend", default = TRUE),
  style = c("facet", "wide"),
  sort = FALSE,
  labels = list(title = NULL, subtitle = NULL, sampleAxis = NULL, countAxis = "counts")
)

\S4method{plotDots}{SingleCellExperiment}(
  object,
  genes,
  perSample = TRUE,
  colMin = -2.5,
  colMax = 2.5,
  dotMin = 0L,
  dotScale = 6L,
  color = getOption(x = "acid.continuous.color", default =
    ggplot2::scale_color_gradient2(low = "orange", mid = "gray75", high = "purple",
    midpoint = 0L)),
  legend = getOption(x = "acid.legend", default = TRUE),
  title = NULL
)

\S4method{plotViolin}{SingleCellExperiment}(
  object,
  genes,
  assay = c("logcounts", "normcounts"),
  perSample = TRUE,
  scale = c("count", "width", "area"),
  color = getOption(x = "acid.discrete.color", default =
    AcidPlots::scale_color_synesthesia_d()),
  legend = getOption(x = "acid.legend", default = TRUE),
  title = NULL
)
}
\arguments{
\item{object}{Object.}

\item{genes}{\code{character} or \code{missing}. Gene identifiers. The function will
automatically match identifiers corresponding to the rownames of the
object, or gene symbols defined in the object.}

\item{assay}{\code{vector(1)}.
Assay name or index position.}

\item{geom}{\code{character(1)}.
Plot type. Uses \code{\link[base:match.arg]{match.arg()}} to pick the type.
Currently supports \code{"dot"} and \code{"violin"}.}

\item{perSample}{\code{logical(1)}.
Visualize the distributions per sample.}

\item{legend}{\code{logical(1)}.
Include plot legend.}

\item{title}{\code{character(1)}.
Title.}

\item{interestingGroups}{\code{character}.
Groups of interest to use for visualization.
Corresponds to factors describing the columns of the object.}

\item{convertGenesToSymbols}{\code{logical(1)}.
Attempt to automatically convert gene identifiers to gene symbols.
Only applies when object contains mappings defined in \code{rowRanges}.}

\item{trans}{\code{character(1)}.
Name of the axis scale transformation to apply.

For more information:\preformatted{help(topic = "scale_x_continuous", package = "ggplot2")
}}

\item{line}{\code{character(1)}.
Include average (median, mean, or geometric mean) line for each interesting
group. Disabled by default and if samples are colored by sample name.}

\item{style}{\code{character(1)}.
Plot style.}

\item{sort}{\code{logical(1)}.
Sort the genes alphabetically.
This setting applies to the gene symbols rather than the gene identifiers
when \code{convertGenesToSymbols} is \code{TRUE}.}

\item{labels}{\code{list}.
ggplot2 labels.
See \code{\link[ggplot2:labs]{ggplot2::labs()}} for details.}

\item{colMin}{\code{numeric(1)}.
Minimum scaled average expression threshold. Everything smaller will be
set to this.}

\item{colMax}{\code{numeric(1)}.
Maximum scaled average expression threshold. Everything larger will be set
to this.}

\item{dotMin}{\code{numeric(1)}.
The fraction of cells at which to draw the smallest dot. All cell groups
with less than this expressing the given gene will have no dot drawn.}

\item{dotScale}{\code{numeric(1)}.
Scale the size of the points, similar to \code{cex}.}

\item{color}{\code{ScaleDiscrete}.
Desired ggplot2 color scale. Must supply discrete values. When set \code{NULL},
the default ggplot2 color palette will be used. If manual color definitions
are desired, we recommend using \code{ggplot2::scale_color_manual()}.

To set the discrete color palette globally, use:\preformatted{options(acid.color.discrete = ggplot2::scale_color_viridis_d())
}}

\item{scale}{\code{character(1)}.
If "area" (default), all violins have the same area (before trimming the
tails). If "count", areas are scaled proportionally to the number of
observartions. If "width", all violins have the same maximum width.
See \code{ggplot2::geom_violin} for details.}

\item{...}{Additional arguments.}
}
\value{
\itemize{
\item \code{style = "facet"}: \code{ggplot} grouped by \code{sampleName}, with
\code{ggplot2::facet_wrap()} applied to panel the samples.
\item \code{style = "wide"}: \code{ggplot} in wide format, with genes on the x-axis.
}
}
\description{
Plot counts
}
\note{
Updated 2022-03-06.
}
\examples{
data(
    RangedSummarizedExperiment,
    SingleCellExperiment_Seurat,
    package = "AcidTest"
)

## SummarizedExperiment ====
object <- RangedSummarizedExperiment
rownames <- head(rownames(object))
print(rownames)
g2s <- AcidGenomes::Gene2Symbol(object)
geneIds <- head(g2s[[1L]])
print(geneIds)
geneNames <- head(g2s[[2L]])
print(geneNames)

## Rownames, gene IDs, and gene names (symbols) are supported.
plotCounts(object, genes = geneIds, style = "facet")
plotCounts(object, genes = geneNames, style = "wide")

## SingleCellExperiment ====
object <- SingleCellExperiment_Seurat

## Plotting with either gene IDs or gene names (symbols) works.
genes <- head(rownames(object), n = 4L)
print(genes)

## Per sample mode enabled.
plotCounts(object, genes = genes, perSample = TRUE)

## Per sample mode disabled.
plotCounts(object, genes = genes, perSample = FALSE)
}
\seealso{
\itemize{
\item \code{Seurat::DotPlot()}.
\item \code{Seurat::VlnPlot()}.
\item \code{Seurat::RidgePlot()}.
\item \code{monocle3::plot_genes_violin()}.
}
}
