% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotCountsPerBiotype-methods.R
\name{plotCountsPerBiotype}
\alias{plotCountsPerBiotype}
\alias{plotCountsPerBroadClass}
\alias{plotCountsPerBiotype,SingleCellExperiment-method}
\alias{plotCountsPerBiotype,SummarizedExperiment-method}
\alias{plotCountsPerBroadClass,SingleCellExperiment-method}
\alias{plotCountsPerBroadClass,SummarizedExperiment-method}
\title{Plot counts per biotype}
\usage{
plotCountsPerBiotype(object, ...)

plotCountsPerBroadClass(object, ...)

\S4method{plotCountsPerBiotype}{SingleCellExperiment}(
  object,
  assay = 1L,
  biotypeCol = "geneBiotype",
  n = 9L,
  interestingGroups = NULL,
  geom = c("violin", "boxplot"),
  trans = c("identity", "log2", "log10"),
  labels = list(title = "Counts per biotype", subtitle = NULL, sampleAxis = NULL,
    countAxis = "counts")
)

\S4method{plotCountsPerBiotype}{SummarizedExperiment}(
  object,
  assay = 1L,
  biotypeCol = "geneBiotype",
  n = 9L,
  interestingGroups = NULL,
  geom = c("violin", "boxplot"),
  trans = c("identity", "log2", "log10"),
  labels = list(title = "Counts per biotype", subtitle = NULL, sampleAxis = NULL,
    countAxis = "counts")
)

\S4method{plotCountsPerBroadClass}{SingleCellExperiment}(
  object,
  ...,
  labels = list(title = "Counts per broad class biotype", subtitle = NULL, sampleAxis =
    NULL, countAxis = "counts")
)

\S4method{plotCountsPerBroadClass}{SummarizedExperiment}(
  object,
  ...,
  labels = list(title = "Counts per broad class biotype", subtitle = NULL, sampleAxis =
    NULL, countAxis = "counts")
)
}
\arguments{
\item{object}{Object.}

\item{assay}{\code{vector(1)}.
Assay name or index position.}

\item{biotypeCol}{\code{character(1)}.
Biotype column name defined in \code{colData()}.}

\item{n}{\code{integer(1)}.
Number to include.}

\item{interestingGroups}{\code{character}.
Groups of interest to use for visualization.
Corresponds to factors describing the columns of the object.}

\item{geom}{\code{character(1)}.
Plot type. Uses \code{\link[base:match.arg]{match.arg()}} internally and defaults to
the first argument in the \code{character} vector.}

\item{trans}{\code{character(1)}.
Name of the axis scale transformation to apply.

For more information:

\if{html}{\out{<div class="sourceCode">}}\preformatted{help(topic = "scale_x_continuous", package = "ggplot2")
}\if{html}{\out{</div>}}}

\item{labels}{\code{list}.
ggplot2 labels.
See \code{ggplot2::labs()} for details.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Plot counts per biotype
}
\note{
Updated 2022-03-04.
}
\examples{
data(
    RangedSummarizedExperiment,
    SingleCellExperiment_splatter,
    package = "AcidTest"
)

## SummarizedExperiment ====
object <- RangedSummarizedExperiment
plotCountsPerBiotype(object)

## SingleCellExperiment ====
object <- SingleCellExperiment_splatter
plotCountsPerBiotype(object)
}
\author{
Michael Steinbaugh, Rory Kirchner
}
