% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCountsPerCell-methods.R
\name{plotCountsPerCell}
\alias{plotCountsPerCell}
\alias{plotCountsPerCell,SingleCellExperiment-method}
\title{Plot counts per cell}
\usage{
plotCountsPerCell(object, ...)

\S4method{plotCountsPerCell}{SingleCellExperiment}(
  object,
  geom = c("histogram", "ecdf", "violin", "ridgeline", "boxplot"),
  interestingGroups = NULL,
  min = 0L,
  max = Inf,
  point = c("none", "inflection", "knee"),
  trans = "log10",
  color = getOption(x = "acid.color.discrete", default = getOption(x =
    "ggplot2.discrete.colour", default = acidplots::scale_color_synesthesia_d())),
  fill = getOption(x = "acid.fill.discrete", default = getOption(x =
    "ggplot2.discrete.fill", default = acidplots::scale_fill_synesthesia_d())),
  title = "Counts per cell"
)
}
\arguments{
\item{object}{Object.}

\item{geom}{\code{character(1)}.
Plot type. Uses \code{\link[base:match.arg]{match.arg()}} internally and defaults to
the first argument in the \code{character} vector.}

\item{interestingGroups}{\code{character}.
Groups of interest to use for visualization.
Corresponds to factors describing the columns of the object.}

\item{min}{\code{numeric(1)}.
Recommended minimum value cutoff.}

\item{max}{\code{numeric(1)}.
Recommended maximum value cutoff.}

\item{point}{\code{character(1)}.
Label either the knee or inflection points per sample.
Requires \code{geom = "ecdf"}.}

\item{trans}{\code{character(1)}.
Name of the axis scale transformation to apply.

For more information:\preformatted{help(topic = "scale_x_continuous", package = "ggplot2")
}}

\item{color}{\code{ScaleDiscrete}.
Desired ggplot2 color scale. Must supply discrete values. When set \code{NULL},
the default ggplot2 color palette will be used. If manual color definitions
are desired, we recommend using \code{ggplot2::scale_color_manual()}.

To set the discrete color palette globally, use:\preformatted{options(acid.color.discrete = ggplot2::scale_color_viridis_d())
}}

\item{fill}{\code{ggproto}/\code{ScaleDiscrete}.
Desired ggplot2 fill scale. Must supply discrete values. When set to
\code{NULL}, the default ggplot2 color palette will be used. If manual color
definitions are desired, we recommend using \code{ggplot2::scale_fill_manual()}.

To set the discrete fill palette globally, use:\preformatted{options(acid.fill.discrete = ggplot2::scale_fill_viridis_d())
}}

\item{title}{\code{character(1)}.
Title.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Plot the disambiguated counts per cell.
}
\details{
"Counts" refer to universal molecular identifier (UMI) counts for
droplet-based scRNA-seq data.
}
\note{
Updated 2019-08-27.
}
\examples{
data(SingleCellExperiment, package = "acidtest")

## SingleCellExperiment ====
object <- SingleCellExperiment
object <- calculateMetrics(object)
plotCountsPerCell(object, geom = "violin")
plotCountsPerCell(object, geom = "ridgeline")
plotCountsPerCell(object, geom = "ecdf")
plotCountsPerCell(object, geom = "histogram")
plotCountsPerCell(object, geom = "boxplot")
}
\seealso{
\code{\link[acidgenerics:plotCountsVsFeatures]{plotCountsVsFeatures()}}.
}
\author{
Michael Steinbaugh, Rory Kirchner
}
