% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotCountsPerFeature-methods.R
\name{plotCountsPerFeature}
\alias{plotCountsPerFeature}
\alias{plotCountsPerFeature,SingleCellExperiment-method}
\alias{plotCountsPerFeature,SummarizedExperiment-method}
\title{Plot counts per feature}
\usage{
plotCountsPerFeature(object, ...)

\S4method{plotCountsPerFeature}{SingleCellExperiment}(object, ...)

\S4method{plotCountsPerFeature}{SummarizedExperiment}(
  object,
  assay = 1L,
  interestingGroups = NULL,
  geom = c("boxplot", "density", "jitter"),
  trans = c("identity", "log2", "log10"),
  labels = list(title = "Counts per feature", subtitle = NULL, sampleAxis = NULL,
    countAxis = "counts"),
  flip = getOption(x = "acid.flip", default = TRUE),
  minMethod = c("absolute", "perRow")
)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{assay}{\code{vector(1)}.
Assay name or index position.}

\item{interestingGroups}{\code{character}.
Groups of interest to use for visualization.
Corresponds to factors describing the columns of the object.}

\item{geom}{\code{character(1)}.
Type of ggplot2 geometric object to use.}

\item{trans}{\code{character(1)}.
Apply a log transformation (e.g. \code{log2(x + 1L)}) to the count matrix prior
to melting, if desired. Use \code{"identity"} to return unmodified (default).}

\item{labels}{\code{list}.
ggplot2 labels.
See \code{\link[ggplot2:labs]{ggplot2::labs()}} for details.}

\item{flip}{\code{logical(1)}.
Flip x and y axes. Recommended for plots containing many samples.}

\item{minMethod}{\code{character(1)}.
Only applies when \code{min} argument is numeric.
Uses \code{match.arg()}.
\itemize{
\item \code{absolute}: Applies hard cutoff to \code{counts} column after the melt
operation. This applies to all counts, not per feature.
\item \code{perRow}: Applies cutoff per row (i.e. gene). Internally, \code{rowSums()}
values are checked against this cutoff threshold prior to the melt
operation.
}}
}
\value{
\code{ggplot}.
}
\description{
Generally, we expect similar count spreads for all genes between samples
unless the library sizes or total RNA expression are different.
}
\section{Functions}{
\itemize{
\item \code{plotCountsPerFeature,SingleCellExperiment-method}: Applies \code{aggregateCellsToSamples()}
calculation to summarize at sample level prior to plotting.\cr
Passes \code{...} to \code{SummarizedExperiment} method.
}}

\note{
Updated 2021-09-10.
}
\examples{
data(
    RangedSummarizedExperiment,
    SingleCellExperiment_splatter,
    package = "AcidTest"
)

## SummarizedExperiment ====
object <- RangedSummarizedExperiment
plotCountsPerFeature(object, geom = "boxplot")
plotCountsPerFeature(object, geom = "density")

## SingleCellExperiment ====
object <- SingleCellExperiment_splatter
plotCountsPerFeature(object)
}
