% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotCountsVsFeatures-methods.R
\name{plotCountsVsFeatures}
\alias{plotCountsVsFeatures}
\alias{plotCountsVsFeatures,SingleCellExperiment-method}
\title{Plot count and feature correlation}
\usage{
plotCountsVsFeatures(object, ...)

\S4method{plotCountsVsFeatures}{SingleCellExperiment}(
  object,
  interestingGroups = NULL,
  trendline = FALSE,
  trans = "log2",
  labels = list(title = "Counts vs. features", subtitle = NULL, x = "counts", y =
    "features")
)
}
\arguments{
\item{object}{Object.}

\item{interestingGroups}{\code{character}.
Groups of interest to use for visualization.
Corresponds to factors describing the columns of the object.}

\item{trendline}{\code{logical(1)}.
Include trendline on plot.}

\item{trans}{\code{character(1)}.
Name of the axis scale transformation to apply.

For more information:\preformatted{help(topic = "scale_x_continuous", package = "ggplot2")
}}

\item{labels}{\code{list}.
ggplot2 labels.
See \code{\link[ggplot2:labs]{ggplot2::labs()}} for details.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Plot the disambiguated counts per cell vs. features (i.e. genes or
transcripts) detected.
}
\details{
"Counts" refer to universal molecular identifier (UMI) counts for
droplet-based scRNA-seq data.
}
\note{
Updated 2022-03-07.
}
\examples{
data(SingleCellExperiment_splatter, package = "AcidTest")

## SingleCellExperiment ====
object <- SingleCellExperiment_splatter
plotCountsVsFeatures(object)
}
\seealso{
\code{\link[AcidGenerics:plotCountsPerCell]{plotCountsPerCell()}}.
}
\author{
Michael Steinbaugh, Rory Kirchner
}
