% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFeaturesDetected-methods.R
\docType{methods}
\name{plotFeaturesDetected}
\alias{plotFeaturesDetected}
\alias{plotFeaturesDetected,SummarizedExperiment-method}
\alias{plotFeaturesDetected,SingleCellExperiment-method}
\title{Plot features detected}
\usage{
plotFeaturesDetected

\S4method{plotFeaturesDetected}{SummarizedExperiment}(object, assay = 1L,
  interestingGroups = NULL, limit = 0L, minCounts = 1L,
  fill = getOption(x = "acid.fill.discrete", default =
  acidplots::scale_fill_synesthesia_d()), flip = getOption(x =
  "acid.flip", default = TRUE), title = "Features detected",
  countsAxisLabel = "features")

\S4method{plotFeaturesDetected}{SingleCellExperiment}(object, assay = 1L,
  interestingGroups = NULL, limit = 0L, minCounts = 1L,
  fill = getOption(x = "acid.fill.discrete", default =
  acidplots::scale_fill_synesthesia_d()), flip = getOption(x =
  "acid.flip", default = TRUE), title = "Features detected",
  countsAxisLabel = "features")
}
\arguments{
\item{object}{Object.}

\item{assay}{\code{vector(1)}.
Name or index of count matrix slotted in
\code{\link[SummarizedExperiment:assays]{assays()}}. When passing in a string, the
name must be defined in
\code{\link[SummarizedExperiment:assayNames]{assayNames()}}.}

\item{interestingGroups}{\code{character}.
Groups of interest that define the samples. If left unset, defaults to
\code{sampleName}.}

\item{limit}{\code{numeric(1)}.
Threshold to denote on the plot, using a dashed line.}

\item{minCounts}{\code{integer(1)}.
Minimum number of counts per gene in the count matrix.}

\item{fill}{\code{ggproto}/\code{ScaleDiscrete}.
Desired ggplot2 fill scale. Must supply discrete values. When set to
\code{NULL}, the default ggplot2 color palette will be used. If manual color
definitions are desired, we recommend using \code{\link[ggplot2:scale_fill_manual]{ggplot2::scale_fill_manual()}}.

To set the discrete fill palette globally, use:\preformatted{options(acid.fill.discrete = ggplot2::scale_fill_viridis_d())
}}

\item{flip}{\code{logical(1)}.
Flip x and y axes. Recommended for plots containing many samples.}

\item{title}{\code{character(1)}.
Plot title.}

\item{countsAxisLabel}{\code{character(1)}.
Counts axis label.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Visualize the number of features (i.e. genes) detected.
}
\examples{
data(
    RangedSummarizedExperiment,
    SingleCellExperiment,
    package = "acidtest"
)
rse <- RangedSummarizedExperiment
sce <- SingleCellExperiment

## SummarizedExperiment ====
plotFeaturesDetected(rse)

## SingleCellExperiment ====
plotFeaturesDetected(sce)
}
