% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotFeaturesPerCell-methods.R
\name{plotFeaturesPerCell}
\alias{plotFeaturesPerCell}
\alias{plotFeaturesPerCell,SingleCellExperiment-method}
\title{Plot features per cell}
\usage{
plotFeaturesPerCell(object, ...)

\S4method{plotFeaturesPerCell}{SingleCellExperiment}(
  object,
  geom = c("histogram", "ecdf", "violin", "ridgeline", "boxplot"),
  interestingGroups = NULL,
  min = 0L,
  max = Inf,
  trans = "log2",
  labels = list(title = "Features per cell", subtitle = NULL, metricAxis = "features",
    otherAxis = NULL)
)
}
\arguments{
\item{object}{Object.}

\item{geom}{\code{character(1)}.
Plot type. Uses \code{\link[base:match.arg]{match.arg()}} internally and defaults to
the first argument in the \code{character} vector.}

\item{interestingGroups}{\code{character}.
Groups of interest to use for visualization.
Corresponds to factors describing the columns of the object.}

\item{min}{\code{numeric(1)}.
Recommended minimum value cutoff.}

\item{max}{\code{numeric(1)}.
Recommended maximum value cutoff.}

\item{trans}{\code{character(1)}.
Name of the axis scale transformation to apply.

For more information:\preformatted{help(topic = "scale_x_continuous", package = "ggplot2")
}}

\item{labels}{\code{list}.
ggplot2 labels.
See \code{\link[ggplot2:labs]{ggplot2::labs()}} for details.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Plot features per cell
}
\note{
Updated 2022-01-19.
}
\examples{
data(SingleCellExperiment_splatter, package = "AcidTest")

## SingleCellExperiment ====
object <- SingleCellExperiment_splatter
object <- calculateMetrics(object)
plotFeaturesPerCell(object)
}
\author{
Michael Steinbaugh, Rory Kirchner
}
