% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotGenderMarkers-methods.R
\name{plotGenderMarkers}
\alias{plotGenderMarkers}
\alias{plotGenderMarkers,SingleCellExperiment-method}
\alias{plotGenderMarkers,SummarizedExperiment-method}
\title{Plot sexually dimorphic gender marker genes}
\usage{
plotGenderMarkers(object, ...)

\S4method{plotGenderMarkers}{SingleCellExperiment}(object, ...)

\S4method{plotGenderMarkers}{SummarizedExperiment}(object, style = "wide", ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Passthrough to \code{plotCounts()}, with the \code{genes} argument
automatically defined.}

\item{style}{\code{character(1)}.
Plot style.}
}
\value{
\code{ggplot}.
}
\description{
This is a convenience function that quickly plots known sexually dimorphic
genes, and overall gene experssion for the X and Y chromosomes.
}
\note{
Currently only \emph{Homo sapiens} and \emph{Mus musculus} genomes are supported.

Updated 2022-03-04.
}
\examples{
data(
    RangedSummarizedExperiment,
    SingleCellExperiment_splatter,
    package = "AcidTest"
)

## SummarizedExperiment ====
object <- RangedSummarizedExperiment
plotGenderMarkers(object)

## SingleCellExperiment ====
object <- SingleCellExperiment_splatter
plotGenderMarkers(object)
}
\seealso{
\code{plotCounts()}.
}
