% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotKnownMarkers-methods.R
\name{plotKnownMarkers}
\alias{plotKnownMarkers}
\alias{plotKnownMarkers,SingleCellExperiment,KnownMarkers-method}
\title{Plot known markers}
\usage{
plotKnownMarkers(object, markers, ...)

\S4method{plotKnownMarkers}{SingleCellExperiment,KnownMarkers}(object, markers, reduction = "UMAP", headerLevel = 2L, ...)
}
\arguments{
\item{object}{Object.}

\item{markers}{Object.}

\item{reduction}{\code{vector(1)}.
Dimension reduction name or index position.}

\item{headerLevel}{\code{integer(1)} (\code{1}-\code{7}).
Markdown header level.}

\item{...}{Passthrough arguments to \code{\link[=plotMarker]{plotMarker()}}.}
}
\value{
Show graphical output.
Invisibly return a \code{ggplot} \code{list}.
}
\description{
Plot known markers
}
\note{
Updated 2023-08-17.
}
\examples{
data(
    KnownMarkers,
    SingleCellExperiment_Seurat,
    package = "AcidTest"
)

## SingleCellExperiment ====
object <- SingleCellExperiment_Seurat
markers <- KnownMarkers
plotKnownMarkers(
    object = object,
    markers = markers,
    reduction = "UMAP"
)
}
