% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotMarker-methods.R
\name{plotMarker}
\alias{plotMarker}
\alias{plotMarker,SingleCellExperiment-method}
\title{Plot cell-type-specific gene markers}
\usage{
plotMarker(object, ...)

\S4method{plotMarker}{SingleCellExperiment}(
  object,
  genes,
  reduction = "UMAP",
  expression = c("mean", "sum"),
  color = getOption(x = "acid.continuous.color", default =
    ggplot2::scale_color_gradient(low = "gray75", high = "purple")),
  pointSize = getOption(x = "acid.point.size", default = 1L),
  pointAlpha = getOption(x = "acid.point.alpha", default = 0.9),
  pointsAsNumbers = getOption(x = "acid.points.as.numbers", default = FALSE),
  label = getOption(x = "acid.label", default = FALSE),
  labelSize = getOption(x = "acid.label.size", default = 6L),
  dark = getOption(x = "acid.dark", default = FALSE),
  legend = getOption(x = "acid.legend", default = TRUE),
  labels = list(title = "auto", subtitle = NULL)
)
}
\arguments{
\item{object}{Object.}

\item{genes}{\code{character}.
Gene identifiers.}

\item{reduction}{\code{vector(1)}.
Dimension reduction name or index position.}

\item{expression}{\code{character(1)}.
Calculation to apply.
Uses \code{\link[base:match.arg]{match.arg()}} internally and defaults to the first
argument in the \code{character} vector.}

\item{color}{\code{ScaleDiscrete}.
Desired ggplot2 color scale. Must supply discrete values. When set \code{NULL},
the default ggplot2 color palette will be used. If manual color definitions
are desired, we recommend using \code{ggplot2::scale_color_manual()}.

To set the discrete color palette globally, use:

\if{html}{\out{<div class="sourceCode">}}\preformatted{options("acid.color.discrete" = ggplot2::scale_color_viridis_d())
}\if{html}{\out{</div>}}}

\item{pointSize}{\code{numeric(1)}.
Point size for dots in the plot.
In the range of 1-3 is generally recommended.}

\item{pointAlpha}{\code{numeric(1)} (\code{0}-\code{1}).
Alpha transparency level.}

\item{pointsAsNumbers}{\code{logical(1)}.
Plot the points as numbers (\code{TRUE}) or dots (\code{FALSE}).}

\item{label}{\code{logical(1)}.
Superimpose sample text labels on the plot.}

\item{labelSize}{\code{integer(1)}.
Size of the text label.}

\item{dark}{\code{logical(1)}.
Plot against a dark background using the \code{acid_theme_light()} ggplot2
theme.}

\item{legend}{\code{logical(1)}.
Include plot legend.}

\item{labels}{\code{list}.
ggplot2 labels.
See \code{ggplot2::labs()} for details.}

\item{...}{Additional arguments.}
}
\value{
Show graphical output.
Invisibly return a \code{ggplot} \code{list}.
}
\description{
Visualize gene markers on a reduced dimension plot (e.g. t-SNE, UMAP).
}
\note{
Updated 2023-04-27.
}
\examples{
requireNamespace("BiocGenerics", quietly = TRUE)
requireNamespace("Matrix", quietly = TRUE)
data(SingleCellExperiment_Seurat, package = "AcidTest")

## SingleCellExperiment ====
object <- SingleCellExperiment_Seurat
sums <- sort(rowSums(counts(object)), decreasing = TRUE)
genes <- names(head(sums, n = 4L))
plotMarker(
    object = object,
    genes = genes,
    reduction = "UMAP"
)
}
\author{
Michael Steinbaugh, Rory Kirchner
}
