% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMitoVsCoding-methods.R
\name{plotMitoVsCoding}
\alias{plotMitoVsCoding}
\alias{plotMitoVsCoding,SingleCellExperiment-method}
\title{Plot mitochondrial vs. coding counts}
\usage{
plotMitoVsCoding(object, ...)

\S4method{plotMitoVsCoding}{SingleCellExperiment}(
  object,
  interestingGroups = NULL,
  trendline = FALSE,
  trans = "log2",
  labels = list(title = "Mito vs. coding", subtitle = NULL, x = "coding", y = "mito")
)
}
\arguments{
\item{object}{Object.}

\item{interestingGroups}{\code{character}.
Groups of interest to use for visualization.
Corresponds to factors describing the columns of the object.}

\item{trendline}{\code{logical(1)}.
Include trendline on plot.}

\item{trans}{\code{character(1)}.
Name of the axis scale transformation to apply.

For more information:\preformatted{help(topic = "scale_x_continuous", package = "ggplot2")
}}

\item{labels}{\code{list}.
ggplot2 labels.
See \code{ggplot2::labs()} for details.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Plot mitochondrial vs. coding counts
}
\note{
Updated 2022-03-07.
}
\examples{
data(SingleCellExperiment_splatter, package = "AcidTest")

## SingleCellExperiment ====
object <- SingleCellExperiment_splatter
try({
    plotMitoVsCoding(object)
})
}
\author{
Michael Steinbaugh, Rory Kirchner
}
