% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotNovelty-methods.R
\name{plotNovelty}
\alias{plotNovelty}
\alias{plotNovelty,SingleCellExperiment-method}
\title{Plot novelty score}
\usage{
plotNovelty(object, ...)

\S4method{plotNovelty}{SingleCellExperiment}(
  object,
  geom = c("histogram", "ecdf", "violin", "ridgeline", "boxplot"),
  interestingGroups = NULL,
  min = 0L,
  trans = "identity",
  labels = list(title = "Novelty", subtitle = NULL, metricAxis = NULL, otherAxis = NULL)
)
}
\arguments{
\item{object}{Object.}

\item{geom}{\code{character(1)}.
Plot type. Uses \code{\link[base:match.arg]{match.arg()}} internally and defaults to
the first argument in the \code{character} vector.}

\item{interestingGroups}{\code{character}.
Groups of interest to use for visualization.
Corresponds to factors describing the columns of the object.}

\item{min}{\code{numeric(1)}.
Recommended minimum value cutoff.}

\item{trans}{\code{character(1)}.
Name of the axis scale transformation to apply.

For more information:

\if{html}{\out{<div class="sourceCode">}}\preformatted{help(topic = "scale_x_continuous", package = "ggplot2")
}\if{html}{\out{</div>}}}

\item{labels}{\code{list}.
ggplot2 labels.
See \code{ggplot2::labs()} for details.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
"Novelty" refers to log10 genes detected per count.
}
\note{
Updated 2022-03-07.
}
\examples{
data(SingleCellExperiment_splatter, package = "AcidTest")

## SingleCellExperiment ====
object <- SingleCellExperiment_splatter
plotNovelty(object)
}
\author{
Michael Steinbaugh
}
