% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotReducedDim-methods.R
\name{plotPCA}
\alias{plotPCA}
\alias{plotPCA,SummarizedExperiment-method}
\title{Principal component analysis plot}
\usage{
\S4method{plotPCA}{SummarizedExperiment}(
  object,
  assay = 1L,
  interestingGroups = NULL,
  ntop = 500L,
  label = getOption(x = "acid.label", default = FALSE),
  pointSize = getOption(x = "acid.point.size", default = 3L),
  labels = list(title = "PCA", subtitle = NULL)
)
}
\arguments{
\item{object}{Object.}

\item{assay}{\code{vector(1)}.
Assay name or index position.}

\item{interestingGroups}{\code{character}.
Groups of interest to use for visualization.
Corresponds to factors describing the columns of the object.}

\item{ntop}{\code{integer(1)} or \code{Inf}.
Number of most variable genes to plot.
Use \code{Inf} to include all genes (\emph{not recommended}).}

\item{label}{\code{logical(1)}.
Superimpose sample text labels on the plot.}

\item{pointSize}{\code{numeric(1)}.
Point size for dots in the plot.
In the range of 1-3 is generally recommended.}

\item{labels}{\code{list}.
ggplot2 labels.
See \code{\link[ggplot2:labs]{ggplot2::labs()}} for details.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Principal component analysis plot
}
\details{
We're using a modified version of the \code{DESeqTransform} method here.\if{html}{\out{<div class="sourceCode r">}}\preformatted{methodFunction(
    f = "plotPCA",
    signature = "DESeqTransform",
    package = "DESeq2"
)
}\if{html}{\out{</div>}}
}
\note{
Updated 2022-03-03.
}
\section{Principal component analysis}{


PCA (Jolliffe, et al., 2002) is a multivariate technique that allows us to
summarize the systematic patterns of variations in the data. PCA takes the
expression levels for genes and transforms it in principal component space,
reducing each sample into one point. Thereby, we can separate samples by
expression variation, and identify potential sample outliers. The PCA plot is
a way to look at how samples are clustering.
}

\section{\code{SingleCellExperiment}}{


The \code{SingleCellExperiment} method that visualizes dimension reduction data
slotted in \code{reducedDims()} is defined in pointillism package.
}

\examples{
data(RangedSummarizedExperiment, package = "AcidTest")

## SummarizedExperiment ====
object <- RangedSummarizedExperiment
plotPCA(object)
}
\references{
Jolliffe, et al., 2002.
}
\seealso{
\itemize{
\item \code{DESeq2::plotPCA()}.
}
}
