% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQuantileHeatmap-methods.R
\docType{methods}
\name{plotQuantileHeatmap}
\alias{plotQuantileHeatmap}
\alias{plotQuantileHeatmap,SummarizedExperiment-method}
\alias{plotQuantileHeatmap,SingleCellExperiment-method}
\title{Quantile heatmap}
\usage{
\S4method{plotQuantileHeatmap}{SummarizedExperiment}(object, assay = 1L,
  interestingGroups = NULL, n = 10L, clusterRows = TRUE,
  clusterCols = TRUE, showRownames = FALSE, showColnames = TRUE,
  treeheightRow = 0L, treeheightCol = 50L, legend = FALSE,
  borderColor = NULL, title = NULL, ...)

\S4method{plotQuantileHeatmap}{SingleCellExperiment}(object, assay = 1L,
  interestingGroups = NULL, n = 10L, clusterRows = TRUE,
  clusterCols = TRUE, showRownames = FALSE, showColnames = TRUE,
  treeheightRow = 0L, treeheightCol = 50L, legend = FALSE,
  borderColor = NULL, title = NULL, ...)
}
\arguments{
\item{object}{Object.}

\item{assay}{\code{vector(1)}.
Name or index of count matrix slotted in
\code{\link[SummarizedExperiment:assays]{assays()}}. When passing in a string, the
name must be defined in
\code{\link[SummarizedExperiment:assayNames]{assayNames()}}.}

\item{interestingGroups}{\code{character}.
Groups of interest that define the samples. If left unset, defaults to
\code{sampleName}.}

\item{n}{\code{integer(1)}.
The number of quantile breaks to create.}

\item{legend}{\code{logical(1)}.
Show the color legend.}

\item{title}{\code{character(1)}.
Plot title.}

\item{...}{Additional arguments.}
}
\value{
\code{pheatmap}.
}
\description{
Quantile heatmap
}
\examples{
data(rse, sce, package = "acidtest")

## SummarizedExperiment ====
plotQuantileHeatmap(rse)

## SingleCellExperiment ====
plotQuantileHeatmap(sce)
}
