% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotStackedBarPlot-methods.R
\name{plotStackedBarPlot}
\alias{plotStackedBarPlot}
\alias{plotStackedBarPlot,SingleCellExperiment-method}
\title{Plot stacked bar plot}
\usage{
plotStackedBarPlot(object, ...)

\S4method{plotStackedBarPlot}{SingleCellExperiment}(
  object,
  absolute = FALSE,
  interestingGroups = NULL,
  labels = NULL
)
}
\arguments{
\item{object}{Object.}

\item{absolute}{\code{logical(1)}.
Return absolute (\code{TRUE}) or relative/proportional (\code{FALSE}) cell count.}

\item{interestingGroups}{\code{character}.
Groups of interest to use for visualization.
Corresponds to factors describing the columns of the object.}

\item{labels}{\code{list}.
ggplot2 labels.
See \code{ggplot2::labs()} for details.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Plot stacked bar plot
}
\note{
Updated 2022-03-05.
}
\examples{
data(SingleCellExperiment_Seurat, package = "AcidTest")

## SingleCellExperiment ====
object <- SingleCellExperiment_Seurat
plotStackedBarPlot(object)
}
