% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUpset-methods.R
\name{plotUpset}
\alias{plotUpset}
\alias{plotUpset,list-method}
\alias{plotUpset,matrix-method}
\alias{plotUpset,data.frame-method}
\alias{plotUpset,DataFrame-method}
\title{UpSet plot}
\usage{
plotUpset(object, ...)

\S4method{plotUpset}{list}(object, ...)

\S4method{plotUpset}{matrix}(object, sets = NULL, orderBySize = c(matrix = TRUE, mainBar = TRUE))

\S4method{plotUpset}{data.frame}(object, sets = NULL, orderBySize = c(matrix = TRUE, mainBar = TRUE))

\S4method{plotUpset}{DataFrame}(object, sets = NULL, orderBySize = c(matrix = TRUE, mainBar = TRUE))
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{sets}{\code{character} or \code{NULL}.
Specific sets to include in plot.
If left \code{NULL}, all sets will be included.}

\item{orderBySize}{\code{logical}.
Whether to order matrix and/or main bar plot by size.}
}
\value{
Graphical output, no return.
}
\description{
S4 wrapper for \code{\link[UpSetR:upset]{UpSetR::upset()}} with improved default aesthetics.
}
\note{
Updated 2020-08-05.
}
\examples{
list <- list(
    a = c("a", "b", "c", "d", "e", "f"),
    b = c("b", "c", "d", "e", "f", "g"),
    c = c("c", "d", "e", "f", "g", "h"),
    d = c("d", "e", "f", "g", "h", "i")
)
plotUpset(list)
}
