% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotZerosVsDepth-methods.R
\name{plotZerosVsDepth}
\alias{plotZerosVsDepth}
\alias{plotZerosVsDepth,SummarizedExperiment-method}
\title{Plot percentage of zeros vs. library depth}
\usage{
plotZerosVsDepth(object, ...)

\S4method{plotZerosVsDepth}{SummarizedExperiment}(
  object,
  assay = 1L,
  interestingGroups = NULL,
  pointAlpha = getOption(x = "acid.point.alpha", default = 0.9),
  pointSize = getOption(x = "acid.point.size", default = 1L),
  labels = list(title = "Zeros vs. depth", subtitle = NULL, x = "library size (depth)", y
    = "dropout rate")
)
}
\arguments{
\item{object}{Object.}

\item{assay}{\code{vector(1)}.
Assay name or index position.}

\item{interestingGroups}{\code{character}.
Groups of interest to use for visualization.
Corresponds to factors describing the columns of the object.}

\item{pointAlpha}{\code{numeric(1)} (\code{0}-\code{1}).
Alpha transparency level.}

\item{pointSize}{\code{numeric(1)}.
Point size for dots in the plot.
In the range of 1-3 is generally recommended.}

\item{labels}{\code{list}.
ggplot2 labels.
See \code{ggplot2::labs()} for details.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Plot percentage of zeros vs. library depth
}
\note{
Updated 2022-11-09.
}
\examples{
data(
    RangedSummarizedExperiment,
    SingleCellExperiment_splatter,
    package = "AcidTest"
)

## SummarizedExperiment ====
object <- RangedSummarizedExperiment
plotZerosVsDepth(object)

## SingleCellExperiment ====
object <- SingleCellExperiment_splatter
plotZerosVsDepth(object)
}
