% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{.data}
\alias{.env}
\alias{!!!}
\alias{!!}
\alias{aes}
\alias{after_stat}
\alias{ggplot}
\alias{ggsave}
\alias{log_breaks}
\alias{plot_annotation}
\alias{plot_layout}
\alias{percent}
\alias{pretty_breaks}
\alias{sym}
\alias{syms}
\alias{theme_set}
\alias{wrap_plots}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ggplot2}{\code{\link[ggplot2:aes]{aes()}}, \code{\link[ggplot2:after_stat]{after_stat()}}, \code{\link[ggplot2:ggplot]{ggplot()}}, \code{\link[ggplot2:ggsave]{ggsave()}}, \code{\link[ggplot2:theme_set]{theme_set()}}}

  \item{patchwork}{\code{\link[patchwork:plot_annotation]{plot_annotation()}}, \code{\link[patchwork:plot_layout]{plot_layout()}}, \code{\link[patchwork:wrap_plots]{wrap_plots()}}}

  \item{rlang}{\code{\link[rlang:!!]{!!()}}, \code{\link[rlang:!!!]{!!!()}}, \code{\link[rlang:.data]{.data}}, \code{\link[rlang:.env]{.env}}, \code{\link[rlang:sym]{sym()}}, \code{\link[rlang:syms]{syms()}}}

  \item{scales}{\code{\link[scales:log_breaks]{log_breaks()}}, \code{\link[scales:percent]{percent()}}, \code{\link[scales:pretty_breaks]{pretty_breaks()}}}
}}

