% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-palettes.R
\name{synesthesia}
\alias{synesthesia}
\alias{synesthesia_pal}
\title{Synesthesia color palette}
\usage{
synesthesia(n = 256L)

synesthesia_pal(n)
}
\description{
Synesthesia color palette
}
\examples{
library(ggplot2)

p <- ggplot(diamonds, aes(carat, price)) + geom_hex()
p + scale_fill_viridis_c()
p + scale_fill_synesthesia_c()

image(matrix(1:400, 20), col = synesthesia())
}
\seealso{
\itemize{
\item \code{colorRamps::matlab.like()}, \code{colorRamps::matlab.like2()}.
\item \code{RColorBrewer::display.brewer.all()}.
\item \code{grDevices::hsv()}, \code{grDevices::rgb()}, \code{grDevices::col2rgb()}.
\item \code{gplots::col2hex()}.
\item \code{ggplot2::continuous_scale()}, \code{ggplot2::discrete_scale()}.
\item \code{scales::gradient_n_pal()}.
\item \code{viridisLite::viridis.map}, \code{viridis::viridis()}, \code{viridis::viridis_pal()}.
\item \code{Seurat::CustomPalette()}, \code{Seurat::PurpleAndYellow()}.
}

Additional URLs:
\itemize{
\item http://www.stat.columbia.edu/~tzheng/files/Rcolor.pdf
\item https://matplotlib.org/tutorials/colors/colormaps.html
\item https://www.nceas.ucsb.edu/~frazier/RSpatialGuides/colorPaletteCheatsheet.pdf
\item http://colorbrewer2.org/
\item http://colorspace.r-forge.r-project.org/
\item http://hclwizard.org/
}
}
