% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-transforms.R
\name{acid_coord_flip}
\alias{acid_coord_flip}
\title{Flip coordinates}
\usage{
acid_coord_flip()
}
\value{
\code{list} of \code{ggproto} objects.
}
\description{
Intelligently flip a plot with discrete data on the X axis.
}
\details{
This function puts the samples that were near the left origin on the X at the
top on the Y axis, making them more human readable.
}
\note{
Updated 2023-08-11.
}
\examples{
library(ggplot2)
g <- ggplot(data = mpg, aes(x = class)) + geom_bar()

## Notice the difference in Y axis sample order.
g + coord_flip()
g + acid_coord_flip()
}
\seealso{
\itemize{
\item \code{coord_flip()}.
}
}
