% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-transforms.R
\name{acid_scale_y_continuous_nopad}
\alias{acid_scale_y_continuous_nopad}
\title{Remove Y axis padding}
\usage{
acid_scale_y_continuous_nopad(
  ...,
  expand = ggplot2::expansion(mult = c(0L, 0.025))
)
}
\arguments{
\item{...}{Other arguments passed to \code{scale_y_continuous()}.}

\item{expand}{\code{numeric}.
Range expansion.
Refer to \code{ggplot2::expansion} for details.}
}
\value{
\code{ggproto}.
}
\description{
Remove Y axis padding
}
\note{
Updated 2021-06-29.
}
\examples{
acid_scale_y_continuous_nopad()
}
