% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-transforms.R
\name{acid_scale_y_continuous_nopad}
\alias{acid_scale_y_continuous_nopad}
\title{Remove Y axis padding}
\usage{
acid_scale_y_continuous_nopad()
}
\value{
\code{ggproto}.
}
\description{
Remove Y axis padding
}
\note{
Updated 2023-08-11.
}
\examples{
library(ggplot2)
g <- ggplot(data = mpg, aes(x = class)) + geom_bar()
## By default, plots contain padding at the top of the graph.
print(g)
## This function will automatically remove the padding.
g + acid_scale_y_continuous_nopad()
}
