% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGlobals.R
\docType{data}
\name{AcidPlotsTestsUrl}
\alias{AcidPlotsTestsUrl}
\title{AcidPlots test data URL}
\format{
An object of class \code{character} of length 1.
}
\usage{
AcidPlotsTestsUrl
}
\description{
AcidPlots test data URL
}
\examples{
AcidPlotsTestsUrl
}
\keyword{internal}
