% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\docType{data}
\name{acidplotsTestsURL}
\alias{acidplotsTestsURL}
\title{acidplots test data URL}
\format{An object of class \code{character} of length 1.}
\usage{
acidplotsTestsURL
}
\description{
acidplots test data URL
}
\examples{
acidplotsTestsURL
}
\keyword{internal}
