% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-autoScale.R
\name{autoScale}
\alias{autoScale}
\alias{autoContinuousColorScale}
\alias{autoContinuousFillScale}
\alias{autoDiscreteColorScale}
\alias{autoDiscreteFillScale}
\title{Automatic color/fill scales to use with ggplot object}
\usage{
autoContinuousColorScale()

autoContinuousFillScale()

autoDiscreteColorScale()

autoDiscreteFillScale()
}
\value{
\code{Scale}/\code{ggproto}/\code{gg}.
}
\description{
Helper functions for setting default colors in Acid Genomics plots.
}
\note{
Updated 2021-09-10.
}
\section{Setting default colors}{


These functions will inherit values defined in global \code{options()}. Either
input a ggplot2 color scale function that returns \code{Scale} class
(e.g. \code{scale_color_synesthesia_c}), or can input \code{"gradient"}, \code{"viridis"}
character strings.
\itemize{
\item \code{autoContinuousColorScale()}: \code{"ggplot2.continuous.colour"}.
\item \code{autoContinuousFillScale()}: \code{"ggplot2.continuous.fill"}.
\item \code{autoDiscreteColorScale()}: \code{"ggplot2.discrete.colour"}.
\item \code{autoDiscreteFillScale()}: \code{"ggplot2.discrete.fill"}.
}
}

\examples{
x <- autoContinuousColorScale()
class(x)
x <- autoContinuousFillScale()
class(x)
x <- autoDiscreteColorScale()
class(x)
x <- autoDiscreteFillScale()
class(x)
}
\seealso{
\itemize{
\item https://ggplot2.tidyverse.org/reference/scale_colour_continuous.html
\item https://ggplot2.tidyverse.org/reference/scale_colour_discrete.html
}
}
