% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-autoScale.R
\name{autoScale}
\alias{autoScale}
\alias{autoContinuousColorScale}
\alias{autoContinuousFillScale}
\alias{autoDiscreteColorScale}
\alias{autoDiscreteFillScale}
\title{Automatic color/fill scales to use with ggplot object}
\usage{
autoContinuousColorScale()

autoContinuousFillScale()

autoDiscreteColorScale()

autoDiscreteFillScale()
}
\value{
\code{Scale}/\code{ggproto}/\code{gg}.
}
\description{
Helper functions for setting default colors in Acid Genomics plots.
}
\note{
Updated 2021-09-10.
}
\examples{
x <- autoContinuousColorScale()
class(x)

x <- autoContinuousFillScale()
class(x)

x <- autoDiscreteColorScale()
class(x)

x <- autoDiscreteFillScale()
class(x)
}
