% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcodeRanksPerSample-methods.R
\name{barcodeRanksPerSample}
\alias{barcodeRanksPerSample}
\alias{barcodeRanksPerSample,SingleCellExperiment-method}
\title{Barcode ranks per sample}
\usage{
barcodeRanksPerSample(object, ...)

\S4method{barcodeRanksPerSample}{SingleCellExperiment}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Passthrough arguments to \code{\link[DropletUtils:barcodeRanks]{DropletUtils::barcodeRanks()}}.}
}
\value{
\code{DataFrameList}.
}
\description{
Barcode ranks per sample
}
\note{
Requires DropletUtils package to be installed.

Updated 2019-11-19.
}
\examples{
data(SingleCellExperiment, package = "acidtest")

## SingleCellExperiment ====
if (requireNamespace("DropletUtils", quietly = TRUE)) {
    object <- SingleCellExperiment
    x <- barcodeRanksPerSample(object)
    names(x)
}
}
\seealso{
\code{\link[DropletUtils:barcodeRanks]{DropletUtils::barcodeRanks()}}.
}
