% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-gradients.R
\name{gradient}
\alias{gradient}
\alias{blueYellow}
\alias{purpleOrange}
\title{Hex color gradient generator}
\usage{
gradient(colors, n)

blueYellow(
  n = 256L,
  palette = c("lightPalette", "darkPalette", "draculaPalette", "macOSLightPalette",
    "macOSDarkPalette", "iOSLightPalette", "iOSDarkPalette")
)

purpleOrange(
  n = 256L,
  palette = c("lightPalette", "darkPalette", "draculaPalette", "macOSLightPalette",
    "macOSDarkPalette", "iOSLightPalette", "iOSDarkPalette")
)
}
\arguments{
\item{colors}{\code{character}.
Color names or hexadecimal values used to define the gradient.}

\item{n}{\code{integer(1)}.
The number of colors (>= 1) to be in the palette.}

\item{palette}{\code{character(1)}.
Color palette name.}
}
\value{
\code{character}.
Hexadecimal colors in RGB space.
}
\description{
Hex color gradient generator
}
\note{
Updated 2022-03-07.
}
\examples{
purpleOrange(n = 3L)
blueYellow(n = 3L)
}
