% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient.R
\name{gradient}
\alias{gradient}
\alias{purpleOrange}
\alias{blueYellow}
\title{Hex color gradient generator}
\usage{
gradient(colors, n)

purpleOrange(n = 256L)

blueYellow(n = 256L)
}
\arguments{
\item{colors}{\code{character}.
Color names or hexadecimal values used to define the gradient.}

\item{n}{\code{integer(1)}.
The number of colors (>= 1) to be in the palette.}
}
\value{
\code{character}.
Hexadecimal colors in RGB space.
}
\description{
Hex color gradient generator
}
\examples{
purpleOrange(n = 3L)
blueYellow(n = 3L)
}
