% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-matchLabels.R
\name{matchLabels}
\alias{matchLabels}
\title{Match ggplot2 labels}
\usage{
matchLabels(labels, argName = "labels")
}
\arguments{
\item{labels}{\code{list}.
User-defined plot labels.
Per element, supports \code{character(1)}., \code{logical(1)}
(for automatic labels), or \code{NULL}.}

\item{argName}{\code{character(1)}.
Argument name defined in \code{formalArgs()} to match against user input.
Defaults to \code{"labels"}.}
}
\value{
\code{list}.
Returns an empty list on \code{NULL} input.
}
\description{
Allow user to define ggplot labels, and populate missing values from the
defaults specified in \code{labels} formal argument.
}
\note{
Updated 2022-01-19.
}
\examples{
fun <- function(object,
                labels = list(
                    "title" = NULL,
                    "x" = "x-axis",
                    "y" = "y-axis"
                )) {
    labels <- matchLabels(labels)
    labels
}
}
\seealso{
\itemize{
\item \code{ggplot2::labs()}.
}
}
