% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-matchLabels.R
\name{matchLabels}
\alias{matchLabels}
\title{Match ggplot2 labels}
\usage{
matchLabels(labels, choices = NULL)
}
\arguments{
\item{labels}{\code{list}.
User-defined plot labels.
Per element, supports \code{character(1)}., \code{logical(1)} (for automatic
labels), or \code{NULL}.}

\item{choices}{\code{list}.
Default plot labels, defined in \code{labels} \code{\link[base:formals]{formals()}}.}
}
\value{
\code{list}.
}
\description{
Allow user to define ggplot labels, and populate missing values from the
defaults specified in \code{labels} formal argument.
}
\details{
Returns an empty list on \code{NULL} input.
}
\note{
Attempting to follow a similar naming convention to
\code{\link[base:match.arg]{match.arg()}} here, using \code{choices} argument.

Updated 2021-08-09.
}
\examples{
matchLabels(
    labels = list(title = "XXX"),
    choices = list(
        title = "AAA",
        subtitle = "BBB"
    )
)
}
\seealso{
\itemize{
\item \code{ggplot2::labs()}.
}
}
