% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGlobals.R
\docType{data}
\name{palettes}
\alias{palettes}
\alias{lightPalette}
\alias{darkPalette}
\alias{draculaPalette}
\alias{iOSLightPalette}
\alias{iOSDarkPalette}
\alias{macOSLightPalette}
\alias{macOSDarkPalette}
\title{Color palettes}
\format{
An object of class \code{character} of length 12.

An object of class \code{character} of length 12.

An object of class \code{character} of length 12.

An object of class \code{character} of length 12.

An object of class \code{character} of length 12.

An object of class \code{character} of length 12.

An object of class \code{character} of length 12.
}
\usage{
lightPalette

darkPalette

draculaPalette

iOSLightPalette

iOSDarkPalette

macOSLightPalette

macOSDarkPalette
}
\value{
\code{character}.
}
\description{
Color palette hexadecimal character vectors.
}
\note{
Updated 2019-09-13.
}
\examples{
lightPalette
}
\seealso{
\itemize{
\item \href{https://developer.apple.com/design/human-interface-guidelines/ios/visual-design/color/}{iOS palette}
\item \href{https://developer.apple.com/design/human-interface-guidelines/macos/visual-design/color/}{macOS palette}
\item \href{https://github.com/dracula/dracula-theme}{Dracula palette}
}
}
\keyword{datasets}
