% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotBarcodeRanks-methods.R
\name{plotBarcodeRanks}
\alias{plotBarcodeRanks}
\alias{plotBarcodeRanks,SingleCellExperiment-method}
\title{Plot barcode ranks}
\usage{
plotBarcodeRanks(object, ...)

\S4method{plotBarcodeRanks}{SingleCellExperiment}(
  object,
  colors = c(fitline = AcidPlots::lightPalette[["blue"]], inflection =
    AcidPlots::lightPalette[["purple"]], knee = AcidPlots::lightPalette[["orange"]]),
  labels = list(title = NULL, subtitle = NULL),
  ...
)
}
\arguments{
\item{object}{Object.}

\item{colors}{\code{character(3)}.
Character vector denoting \code{fitline}, \code{inflection}, and \code{knee} point colors.
Must pass in color names or hexadecimal values.}

\item{labels}{\code{list}.
ggplot2 labels.
See \code{ggplot2::labs()} for details.}

\item{...}{Passthrough to \code{\link[=barcodeRanksPerSample]{barcodeRanksPerSample()}}.}
}
\value{
\code{ggplot} grid.
}
\description{
Plot barcode ranks
}
\note{
Requires DropletUtils package to be installed.

Updated 2022-01-19.
}
\examples{
data(SingleCellExperiment_splatter, package = "AcidTest")

## SingleCellExperiment ====
if (requireNamespace("DropletUtils", quietly = TRUE)) {
    object <- SingleCellExperiment_splatter
    plotBarcodeRanks(object)
}
}
