% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotCellCounts-methods.R
\name{plotCellCounts}
\alias{plotCellCounts}
\alias{plotCellCounts,SingleCellExperiment-method}
\title{Plot cell counts}
\usage{
plotCellCounts(object, ...)

\S4method{plotCellCounts}{SingleCellExperiment}(
  object,
  interestingGroups = NULL,
  fill = getOption(x = "acid.fill.discrete", default = getOption(x =
    "ggplot2.discrete.fill", default = AcidPlots::scale_fill_synesthesia_d())),
  labels = list(title = "Cell counts", subtitle = NULL, x = NULL, y = "cells")
)
}
\arguments{
\item{object}{Object.}

\item{interestingGroups}{\code{character}.
Groups of interest to use for visualization.
Corresponds to factors describing the columns of the object.}

\item{fill}{\code{ggproto}/\code{ScaleDiscrete}.
Desired ggplot2 fill scale. Must supply discrete values. When set to
\code{NULL}, the default ggplot2 color palette will be used. If manual color
definitions are desired, we recommend using \code{ggplot2::scale_fill_manual()}.

To set the discrete fill palette globally, use:\preformatted{options(acid.fill.discrete = ggplot2::scale_fill_viridis_d())
}}

\item{labels}{\code{list}.
ggplot2 labels.
See \code{\link[ggplot2:labs]{ggplot2::labs()}} for details.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Plot cell counts
}
\note{
Updated 2019-12-09.
}
\examples{
data(SingleCellExperiment, package = "AcidTest")

## SingleCellExperiment ====
object <- SingleCellExperiment
plotCellCounts(object)
}
\author{
Michael Steinbaugh, Rory Kirchner
}
