% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotCorrelation-methods.R
\name{plotCorrelation}
\alias{plotCorrelation}
\alias{plotCorrelation,matrix-method}
\alias{plotCorrelation,Matrix-method}
\alias{plotCorrelation,data.frame-method}
\alias{plotCorrelation,DataFrame-method}
\alias{plotCorrelation,SummarizedExperiment-method}
\title{Correlation X-Y scatterplot}
\usage{
plotCorrelation(object, ...)

\S4method{plotCorrelation}{matrix}(
  object,
  xCol,
  yCol,
  label = FALSE,
  labels = NULL,
  trans = c("identity", "log2", "log10"),
  colors = list(dots = "black", line = "black", se = "gray")
)

\S4method{plotCorrelation}{Matrix}(
  object,
  xCol,
  yCol,
  label = FALSE,
  labels = NULL,
  trans = c("identity", "log2", "log10"),
  colors = list(dots = "black", line = "black", se = "gray")
)

\S4method{plotCorrelation}{data.frame}(
  object,
  xCol,
  yCol,
  label = FALSE,
  labels = NULL,
  trans = c("identity", "log2", "log10"),
  colors = list(dots = "black", line = "black", se = "gray")
)

\S4method{plotCorrelation}{DataFrame}(
  object,
  xCol,
  yCol,
  label = FALSE,
  labels = NULL,
  trans = c("identity", "log2", "log10"),
  colors = list(dots = "black", line = "black", se = "gray")
)

\S4method{plotCorrelation}{SummarizedExperiment}(object, assay = 1L, ...)
}
\arguments{
\item{object}{Object.}

\item{xCol, yCol}{\code{character(1)} or \code{integer(1)}.
X and Y column name or position.}

\item{label}{\code{logical(1)}.
Superimpose sample text labels on the plot.}

\item{labels}{\code{list}.
ggplot2 labels.
See \code{\link[ggplot2:labs]{ggplot2::labs()}} for details.}

\item{trans}{\code{character(1)}.
Name of the axis scale transformation to apply.

For more information:\preformatted{help(topic = "scale_x_continuous", package = "ggplot2")
}}

\item{colors}{\code{list(3)}.
Named list defining the colors of \code{dots}, \code{line}, and \code{se}, for confidence
interval standard error.}

\item{assay}{\code{vector(1)}.
Assay name or index position.}

\item{...}{Additional arguments.}
}
\value{
Plot.
}
\description{
Correlation X-Y scatterplot
}
\note{
Updated 2021-02-09.
}
\section{Correlation coefficient calculations}{


Correlation coefficient calcluations are generated by
\code{ggpmisc::stat_poly_eq()}. Refer to the
\href{https://github.com/aphalo/ggpmisc}{ggpmisc GitHub repo} for details.
}

\examples{
data(RangedSummarizedExperiment, package = "AcidTest")

## SummarizedExperiment ====
object <- RangedSummarizedExperiment
plotCorrelation(object, xCol = 1L, yCol = 2L)
}
