% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotCorrelation-methods.R
\name{plotCorrelation}
\alias{plotCorrelation}
\alias{plotCorrelation,matrix-method}
\alias{plotCorrelation,Matrix-method}
\alias{plotCorrelation,data.frame-method}
\alias{plotCorrelation,DataFrame-method}
\alias{plotCorrelation,SummarizedExperiment-method}
\title{Correlation X-Y scatterplot}
\usage{
plotCorrelation(object, ...)

\S4method{plotCorrelation}{matrix}(
  object,
  xCol,
  yCol,
  label = FALSE,
  labels = NULL,
  trans = c("log10", "log2", "identity"),
  r2 = TRUE,
  se = TRUE,
  colors = list(dots = "black", line = "black", se = "gray")
)

\S4method{plotCorrelation}{Matrix}(
  object,
  xCol,
  yCol,
  label = FALSE,
  labels = NULL,
  trans = c("log10", "log2", "identity"),
  r2 = TRUE,
  se = TRUE,
  colors = list(dots = "black", line = "black", se = "gray")
)

\S4method{plotCorrelation}{data.frame}(object, labelCol = NULL, ...)

\S4method{plotCorrelation}{DataFrame}(object, labelCol = NULL, ...)

\S4method{plotCorrelation}{SummarizedExperiment}(object, assay = 1L, ...)
}
\arguments{
\item{object}{Object.}

\item{xCol, yCol}{\code{character(1)} or \code{integer(1)}.
X and Y column name or position.}

\item{label}{\code{logical(1)}.
Superimpose sample text labels on the plot.}

\item{labels}{\code{list}.
ggplot2 labels.
See \code{\link[ggplot2:labs]{ggplot2::labs()}} for details.}

\item{trans}{\code{character(1)}.
Name of the axis scale transformation to apply.

For more information:\preformatted{help(topic = "scale_x_continuous", package = "ggplot2")
}}

\item{r2}{\code{logical(1)}.
Show information on the \code{lm} fit?
This includes the equation, and the coefficient of determination (R^2).
Refer to \code{ggpmisc::stat_poly_eq} for details.}

\item{se}{\code{logical(1)}.
Display confidence interval around the \code{lm} fit line?
Refer to \code{ggplot2::geom_smooth} for details.}

\item{colors}{\code{list(3)}.
Named list defining the colors of \code{dots}, \code{line}, and \code{se}, for confidence
interval standard error.}

\item{labelCol}{\code{character(1)} or NULL.
For \code{data.frame} method, which column should be used to label points
on the plot?}

\item{...}{Additional arguments.}

\item{assay}{\code{vector(1)}.
Assay name or index position.}
}
\value{
Plot.
}
\description{
Correlation X-Y scatterplot
}
\note{
Updated 2021-08-09.
}
\section{Correlation coefficient calculations}{


Correlation coefficient calcluations are generated by
\code{ggpmisc::stat_poly_eq()}. Refer to the
\href{https://github.com/aphalo/ggpmisc}{ggpmisc GitHub repo} for details.
}

\examples{
data(RangedSummarizedExperiment, package = "AcidTest")

## SummarizedExperiment ====
object <- RangedSummarizedExperiment
plotCorrelation(
    object = object,
    xCol = 1L,
    yCol = 2L,
    trans = "identity"
)
}
\seealso{
\itemize{
\item \code{ggpmisc::stat_poly_eq}.
\item \code{ggplot2::geom_smooth}.
}
}
