% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotCorrelation-methods.R
\name{plotCorrelation}
\alias{plotCorrelation}
\title{Correlation X-Y scatterplot}
\usage{
plotCorrelation(object, ...)
}
\arguments{
\item{...}{Additional arguments.}
}
\value{
Plot.
}
\description{
Correlation X-Y scatterplot
}
\note{
Updated 2021-02-09.
}
\examples{
data(
    RangedSummarizedExperiment,
    package = "AcidTest"
)

## SummarizedExperiment ====
object <- RangedSummarizedExperiment
plotCorrelationHeatmap(object)
}
