% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCorrelationHeatmap-methods.R
\name{plotCorrelationHeatmap}
\alias{plotCorrelationHeatmap}
\alias{plotCorrelationHeatmap,SummarizedExperiment-method}
\alias{plotCorrelationHeatmap,SingleCellExperiment-method}
\title{Correlation heatmap}
\usage{
plotCorrelationHeatmap(object, ...)

\S4method{plotCorrelationHeatmap}{SummarizedExperiment}(
  object,
  assay = 1L,
  interestingGroups = NULL,
  method = c("pearson", "kendall", "spearman"),
  clusteringMethod = "ward.D2",
  showRownames = TRUE,
  showColnames = TRUE,
  treeheightRow = 0L,
  treeheightCol = 50L,
  color = getOption(x = "acid.heatmap.color", default = acidplots::synesthesia),
  legendColor = getOption(x = "acid.heatmap.color", default = acidplots::synesthesia),
  borderColor = NULL,
  title = TRUE,
  ...
)

\S4method{plotCorrelationHeatmap}{SingleCellExperiment}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{assay}{\code{vector(1)}.
Assay name or index position.}

\item{interestingGroups}{\code{character}.
Groups of interest to use for visualization.
Corresponds to factors describing the columns of the object.}

\item{method}{\code{character(1)}.
Correlation coefficient (or covariance) method to be computed.
Defaults to pearson, but spearman or kendall can also be used.
Refer to the \code{\link[stats:cor]{cor()}} documentation for details.}

\item{clusteringMethod}{\code{character(1)}.
Clustering method. Accepts the same values as \code{\link[stats:hclust]{hclust()}}.}

\item{showRownames}{\code{logical(1)}.
Show row or column names.}

\item{showColnames}{\code{logical(1)}.
Show row or column names.}

\item{treeheightRow}{\code{integer(1)}.
Size of the row and column dendrograms. Use \code{0} to disable.}

\item{treeheightCol}{\code{integer(1)}.
Size of the row and column dendrograms. Use \code{0} to disable.}

\item{color}{\verb{function}, \code{character}, or \code{NULL}.
Hexadecimal color function or values to use for plot.

We generally recommend these hexadecimal functions from the viridis
package, in addition to our \code{\link[=synesthesia]{synesthesia()}} palette:
\itemize{
\item \code{viridis::viridis()}.
\item \code{viridis::inferno()}.
\item \code{viridis::magma()}.
\item \code{viridis::plasma()}.
}

Alternatively, colors can be defined manually using hexadecimal values
(e.g. \code{c("#FF0000", "#0000FF")}), but this is not generally recommended.
Refer to the RColorBrewer package for hexadecimal color palettes that may
be suitable. If set \code{NULL}, will use the default pheatmap colors.}

\item{legendColor}{\verb{function} or \code{NULL}.
Hexadecimal color function to use for legend labels. Note that hexadecimal
values are not supported. If set \code{NULL}, will use the default pheatmap
colors.}

\item{borderColor}{\code{character(1)} or \code{NULL}.
Border color.}

\item{title}{\code{character(1)}.
Title.}

\item{...}{Additional arguments.}
}
\value{
\code{pheatmap}.
}
\description{
Correlation heatmap
}
\section{Functions}{
\itemize{
\item \code{plotCorrelationHeatmap,SingleCellExperiment-method}: Applies \code{\link[=pseudobulk]{pseudobulk()}} calculation to
average gene expression at sample level prior to plotting.\cr
Passes \code{...} to \code{SummarizedExperiment} method.
}}

\note{
Updated 2019-09-15.
}
\examples{
data(
    RangedSummarizedExperiment,
    SingleCellExperiment,
    package = "acidtest"
)

## SummarizedExperiment ====
object <- RangedSummarizedExperiment
plotCorrelationHeatmap(object)

## SingleCellExperiment ====
object <- SingleCellExperiment
plotCorrelationHeatmap(object)
}
