% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotCounts-methods.R
\name{plotCounts}
\alias{plotCounts}
\alias{plotDots}
\alias{plotViolin}
\alias{plotCounts,SingleCellExperiment-method}
\alias{plotCounts,SummarizedExperiment-method}
\title{Plot counts}
\usage{
plotCounts(object, ...)

plotDots(object, ...)

plotViolin(object, ...)

\S4method{plotCounts}{SingleCellExperiment}(
  object,
  genes,
  assay = c("logcounts", "normcounts"),
  geom = c("violin", "dot"),
  perSample = TRUE,
  legend = getOption(x = "acid.legend", default = TRUE),
  title = NULL
)

\S4method{plotCounts}{SummarizedExperiment}(
  object,
  genes,
  assay = 1L,
  interestingGroups = NULL,
  convertGenesToSymbols = TRUE,
  geom = c("point", "violin", "boxplot", "bar"),
  trans = c("identity", "log2", "log10"),
  line = c("none", "median", "mean", "geometricMean"),
  legend = getOption(x = "acid.legend", default = TRUE),
  style = c("facet", "wide"),
  sort = FALSE,
  labels = list(title = NULL, subtitle = NULL, sampleAxis = NULL, countAxis = "counts")
)
}
\arguments{
\item{object}{Object.}

\item{genes}{\code{character} or \code{missing}. Gene identifiers. The function will
automatically match identifiers corresponding to the rownames of the
object, or gene symbols defined in the object.}

\item{assay}{\code{vector(1)}.
Assay name or index position.}

\item{geom}{\code{character(1)}.
Plot type. Uses \code{\link[base:match.arg]{match.arg()}} internally and defaults to
the first argument in the \code{character} vector.}

\item{perSample}{\code{logical(1)}.
Visualize the distributions per sample.}

\item{legend}{\code{logical(1)}.
Include plot legend.}

\item{title}{\code{character(1)}.
Title.}

\item{interestingGroups}{\code{character}.
Groups of interest to use for visualization.
Corresponds to factors describing the columns of the object.}

\item{convertGenesToSymbols}{\code{logical(1)}.
Attempt to automatically convert gene identifiers to gene symbols.
Only applies when object contains mappings defined in \code{rowRanges}.}

\item{trans}{\code{character(1)}.
Name of the axis scale transformation to apply.

For more information:\preformatted{help(topic = "scale_x_continuous", package = "ggplot2")
}}

\item{line}{\code{character(1)}.
Include average (median, mean, or geometric mean) line for each interesting
group. Disabled by default and if samples are colored by sample name.}

\item{style}{\code{character(1)}.
Plot style.}

\item{sort}{\code{logical(1)}.
Sort the genes alphabetically.
This setting applies to the gene symbols rather than the gene identifiers
when \code{convertGenesToSymbols} is \code{TRUE}.}

\item{labels}{\code{list}.
ggplot2 labels.
See \code{\link[ggplot2:labs]{ggplot2::labs()}} for details.}

\item{...}{Additional arguments.}
}
\value{
\itemize{
\item \code{style = "facet"}: \code{ggplot} grouped by \code{sampleName}, with
\code{ggplot2::facet_wrap()} applied to panel the samples.
\item \code{style = "wide"}: \code{ggplot} in wide format, with genes on the x-axis.
}
}
\description{
Plot counts
}
\note{
Updated 2022-01-19.
}
\examples{
data(
    RangedSummarizedExperiment,
    SingleCellExperiment_Seurat,
    package = "AcidTest"
)

## SummarizedExperiment ====
object <- RangedSummarizedExperiment
rownames <- head(rownames(object))
print(rownames)
g2s <- basejump::Gene2Symbol(object)
geneIds <- head(g2s[[1L]])
print(geneIds)
geneNames <- head(g2s[[2L]])
print(geneNames)

## Rownames, gene IDs, and gene names (symbols) are supported.
plotCounts(object, genes = geneIds, style = "facet")
plotCounts(object, genes = geneNames, style = "wide")

## SingleCellExperiment ====
object <- SingleCellExperiment_Seurat

## Plotting with either gene IDs or gene names (symbols) works.
genes <- head(rownames(object), n = 4L)
print(genes)

## Per sample mode enabled.
plotCounts(object, genes = genes, perSample = TRUE)

## Per sample mode disabled.
plotCounts(object, genes = genes, perSample = FALSE)
}
\seealso{
\code{DESeq2::plotCounts()}.
}
