% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/plotCountsCorrelation-methods.R
\name{plotCountsCorrelation}
\alias{plotCountsCorrelation}
\alias{plotCountsCorrelation,matrix,matrix-method}
\title{Plot counts correlation}
\usage{
plotCountsCorrelation(x, y, ...)

\S4method{plotCountsCorrelation}{matrix,matrix}(
  x,
  y,
  i = NULL,
  j = NULL,
  labels = list(title = NULL, subtitle = NULL, color = NULL, x = NULL, y = "counts"),
  .xname = getNameInParent(x),
  .yname = getNameInParent(y)
)
}
\arguments{
\item{x}{Object.}

\item{y}{Object.}

\item{i}{Indices specifying elements to extract or replace. Indices are \code{numeric} or
\code{character} vectors, empty (\code{missing}), or \code{NULL}.

For more information:

\if{html}{\out{<div class="sourceCode">}}\preformatted{help(topic = "Extract", package = "base")
}\if{html}{\out{</div>}}}

\item{j}{Indices specifying elements to extract or replace. Indices are \code{numeric} or
\code{character} vectors, empty (\code{missing}), or \code{NULL}.

For more information:

\if{html}{\out{<div class="sourceCode">}}\preformatted{help(topic = "Extract", package = "base")
}\if{html}{\out{</div>}}}

\item{labels}{\code{list}.
ggplot2 labels.
See \code{ggplot2::labs()} for details.}

\item{.xname}{Name of object defined in \code{x}.
\emph{Not intended to be used directly.}}

\item{.yname}{Name of object defined in \code{y}.
\emph{Not intended to be used directly.}}

\item{...}{Additional arguments.}
}
\value{
Graphical output.
}
\description{
Compare feature expression correlation across two data sets.
}
\note{
Updated 2021-09-10.
}
\examples{
data(RangedSummarizedExperiment, package = "AcidTest")

## matrix ====
x <- SummarizedExperiment::assay(RangedSummarizedExperiment)
x <- x[seq_len(4L), seq_len(2L)]
y <- x * 2L
plotCountsCorrelation(x, y)
}
