% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCountsCorrelation-methods.R
\docType{methods}
\name{plotCountsCorrelation}
\alias{plotCountsCorrelation}
\alias{plotCountsCorrelation,matrix,matrix-method}
\title{Plot counts correlation}
\usage{
plotCountsCorrelation(x, y, ...)

\S4method{plotCountsCorrelation}{matrix,matrix}(x, y, i = NULL,
  j = NULL, xTitle = getNameInParent(x), yTitle = getNameInParent(y),
  labs = list(colour = NULL, x = NULL, y = NULL, title = NULL, subtitle =
  NULL))
}
\arguments{
\item{x}{Object.}

\item{y}{Object.}

\item{i}{
    indices specifying elements to extract or replace.  Indices are
    \code{numeric} or \code{character} vectors or empty (missing) or
    \code{NULL}.  Numeric values are coerced to integer as by
    \code{\link{as.integer}} (and hence truncated towards zero).
    Character vectors will be matched to the \code{\link{names}} of the
    object (or for matrices/arrays, the \code{\link{dimnames}}):
    see \sQuote{Character indices} below for further details.

    For \code{[}-indexing only: \code{i}, \code{j}, \code{\dots} can be
    logical vectors, indicating elements/slices to select.  Such vectors
    are recycled if necessary to match the corresponding extent.
    \code{i}, \code{j}, \code{\dots} can also be negative integers,
    indicating elements/slices to leave out of the selection.

    When indexing arrays by \code{[} a single argument \code{i} can be a
    matrix with as many columns as there are dimensions of \code{x}; the
    result is then a vector with elements corresponding to the sets of
    indices in each row of \code{i}.

    An index value of \code{NULL} is treated as if it were \code{integer(0)}.
  }

\item{j}{
    indices specifying elements to extract or replace.  Indices are
    \code{numeric} or \code{character} vectors or empty (missing) or
    \code{NULL}.  Numeric values are coerced to integer as by
    \code{\link{as.integer}} (and hence truncated towards zero).
    Character vectors will be matched to the \code{\link{names}} of the
    object (or for matrices/arrays, the \code{\link{dimnames}}):
    see \sQuote{Character indices} below for further details.

    For \code{[}-indexing only: \code{i}, \code{j}, \code{\dots} can be
    logical vectors, indicating elements/slices to select.  Such vectors
    are recycled if necessary to match the corresponding extent.
    \code{i}, \code{j}, \code{\dots} can also be negative integers,
    indicating elements/slices to leave out of the selection.

    When indexing arrays by \code{[} a single argument \code{i} can be a
    matrix with as many columns as there are dimensions of \code{x}; the
    result is then a vector with elements corresponding to the sets of
    indices in each row of \code{i}.

    An index value of \code{NULL} is treated as if it were \code{integer(0)}.
  }

\item{xTitle}{\code{character(1)}.
Name of object defined in \code{x}.}

\item{yTitle}{\code{character(1)}.
Name of object defined in \code{y}.}

\item{labs}{\code{list}.
ggplot2 labels. See \code{\link[ggplot2:labs]{ggplot2::labs()}} for details.}

\item{...}{Additional arguments.}
}
\value{
Graphical output.
}
\description{
Compare feature expression correlation across two data sets.
}
\note{
Updated 2019-08-27.
}
\examples{
data(RangedSummarizedExperiment, package = "acidtest")
rse <- RangedSummarizedExperiment

## matrix ====
x <- SummarizedExperiment::assay(rse)
x <- x[seq_len(4L), seq_len(2L)]
y <- x * 2L
plotCountsCorrelation(x, y)
}
