% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/plotCountsCorrelationHeatmap-methods.R
\name{plotCountsCorrelationHeatmap}
\alias{plotCountsCorrelationHeatmap}
\alias{plotCountsCorrelationHeatmap,matrix,matrix-method}
\title{Plot counts correlation heatmap}
\usage{
plotCountsCorrelationHeatmap(x, y, ...)

\S4method{plotCountsCorrelationHeatmap}{matrix,matrix}(x, y, i = NULL, j = NULL, method = "pearson", ...)
}
\arguments{
\item{x}{Object.}

\item{y}{Object.}

\item{i}{Indices specifying elements to extract or replace. Indices are \code{numeric} or
\code{character} vectors, empty (\code{missing}), or \code{NULL}.

For more information:\preformatted{help(topic = "Extract", package = "base")
}}

\item{j}{Indices specifying elements to extract or replace. Indices are \code{numeric} or
\code{character} vectors, empty (\code{missing}), or \code{NULL}.

For more information:\preformatted{help(topic = "Extract", package = "base")
}}

\item{method}{\code{character(1)}.
Correlation method.
Refer to \code{stats::cor()} for details.}

\item{...}{Passthrough arguments to \code{plotHeatmap()}.}
}
\value{
Graphical output.
}
\description{
Compare feature expression correlation across two data sets.
}
\note{
Updated 2021-02-08.
}
\examples{
data(RangedSummarizedExperiment, package = "AcidTest")

## matrix ====
x <- basejump::assay(RangedSummarizedExperiment)
y <- x + 1L
plotCountsCorrelationHeatmap(x, y)
}
