% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCountsPerBiotype-methods.R
\docType{methods}
\name{plotCountsPerBiotype}
\alias{plotCountsPerBiotype}
\alias{plotCountsPerBiotype,SummarizedExperiment-method}
\alias{plotCountsPerBiotype,SingleCellExperiment-method}
\title{Plot counts per biotype}
\usage{
plotCountsPerBiotype(object, ...)

\S4method{plotCountsPerBiotype}{SummarizedExperiment}(object, assay = 1L,
  n = 9L, interestingGroups = NULL, trans = c("identity", "log2",
  "log10"), fill = getOption(x = "acid.fill.discrete", default =
  acidplots::scale_fill_synesthesia_d()), labels = list(title =
  "Counts per biotype", subtitle = NULL, samplesAxis = NULL, countsAxis =
  "counts"))

\S4method{plotCountsPerBiotype}{SingleCellExperiment}(object, assay = 1L,
  n = 9L, interestingGroups = NULL, trans = c("identity", "log2",
  "log10"), fill = getOption(x = "acid.fill.discrete", default =
  acidplots::scale_fill_synesthesia_d()), labels = list(title =
  "Counts per biotype", subtitle = NULL, samplesAxis = NULL, countsAxis =
  "counts"))
}
\arguments{
\item{object}{Object.}

\item{assay}{\code{vector(1)}.
Assay name or index position.}

\item{n}{\code{integer(1)}.
Number to include.}

\item{interestingGroups}{\code{character}.
Groups of interest to use for visualization.
Corresponds to factors describing the columns of the object.}

\item{trans}{\code{character(1)}.
Name of the axis scale transformation to apply.

For more information:\preformatted{help(topic = "scale_x_continuous", package = "ggplot2")
}}

\item{fill}{\code{ggproto}/\code{ScaleDiscrete}.
Desired ggplot2 fill scale. Must supply discrete values. When set to
\code{NULL}, the default ggplot2 color palette will be used. If manual color
definitions are desired, we recommend using \code{\link[ggplot2:scale_fill_manual]{ggplot2::scale_fill_manual()}}.

To set the discrete fill palette globally, use:\preformatted{options(acid.fill.discrete = ggplot2::scale_fill_viridis_d())
}}

\item{labels}{\code{list}.
ggplot2 labels.
See \code{\link[ggplot2:labs]{ggplot2::labs()}} for details.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Plot counts per biotype
}
\note{
Updated 2019-09-16.
}
\examples{
data(
    RangedSummarizedExperiment,
    SingleCellExperiment,
    package = "acidtest"
)

## SummarizedExperiment ====
object <- RangedSummarizedExperiment
plotCountsPerBiotype(object)

## SingleCellExperiment ====
object <- SingleCellExperiment
plotCountsPerBiotype(object)
}
\author{
Michael Steinbaugh, Rory Kirchner
}
