% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCountsPerBroadClass-methods.R
\docType{methods}
\name{plotCountsPerBroadClass}
\alias{plotCountsPerBroadClass}
\alias{plotCountsPerBroadClass,SummarizedExperiment-method}
\title{Plot counts per broad class definition}
\usage{
plotCountsPerBroadClass(object, ...)

\S4method{plotCountsPerBroadClass}{SummarizedExperiment}(object,
  assay = 1L, interestingGroups = NULL, trans = c("identity", "log2",
  "log10"), fill = getOption(x = "acid.fill.discrete", default =
  acidplots::scale_fill_synesthesia_d()), countsAxisLabel = "counts",
  title = "Counts per broad class")
}
\arguments{
\item{object}{Object.}

\item{assay}{\code{vector(1)}.
Name or index of count matrix slotted in
\code{\link[SummarizedExperiment:assays]{assays()}}. When passing in a string, the
name must be defined in
\code{\link[SummarizedExperiment:assayNames]{assayNames()}}.}

\item{interestingGroups}{\code{character}.
Groups of interest that define the samples. If left unset, defaults to
\code{sampleName}.}

\item{trans}{\code{character(1)}.
Name of the axis scale transformation to apply.

For more information:\preformatted{help(topic = "scale_x_continuous", package = "ggplot2")
}}

\item{fill}{\code{ggproto}/\code{ScaleDiscrete}.
Desired ggplot2 fill scale. Must supply discrete values. When set to
\code{NULL}, the default ggplot2 color palette will be used. If manual color
definitions are desired, we recommend using \code{\link[ggplot2:scale_fill_manual]{ggplot2::scale_fill_manual()}}.

To set the discrete fill palette globally, use:\preformatted{options(acid.fill.discrete = ggplot2::scale_fill_viridis_d())
}}

\item{countsAxisLabel}{\code{character(1)}.
Counts axis label.}

\item{title}{\code{character(1)}.
Plot title.}

\item{...}{Additional arguments.}
}
\value{
Plot.
}
\description{
Plot counts per broad class definition
}
\examples{
data(
    RangedSummarizedExperiment,
    SingleCellExperiment,
    package = "acidtest"
)
rse <- RangedSummarizedExperiment
sce <- SingleCellExperiment

## SummarizedExperiment ====
plotCountsPerBroadClass(rse)

## SingleCellExperiment ====
plotCountsPerBroadClass(sce)
}
\author{
Michael Steinbaugh, Rory Kirchner
}
