% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/plotCountsPerBroadClass-methods.R
\name{plotCountsPerBroadClass}
\alias{plotCountsPerBroadClass}
\alias{plotCountsPerBroadClass,SingleCellExperiment-method}
\alias{plotCountsPerBroadClass,SummarizedExperiment-method}
\title{Plot counts per broad class definition}
\usage{
plotCountsPerBroadClass(object, ...)

\S4method{plotCountsPerBroadClass}{SingleCellExperiment}(
  object,
  ...,
  labels = list(title = "Counts per broad class biotype", subtitle = NULL, sampleAxis =
    NULL, countAxis = "counts")
)

\S4method{plotCountsPerBroadClass}{SummarizedExperiment}(
  object,
  ...,
  labels = list(title = "Counts per broad class biotype", subtitle = NULL, sampleAxis =
    NULL, countAxis = "counts")
)
}
\arguments{
\item{object}{Object.}

\item{...}{Passthrough arguments to \code{plotCountsPerBiotype()}.}

\item{labels}{\code{list}.
ggplot2 labels.
See \code{\link[ggplot2:labs]{ggplot2::labs()}} for details.}
}
\value{
Plot.
}
\description{
Plot counts per broad class definition
}
\note{
Updated 2021-09-09.
}
\examples{
data(
    RangedSummarizedExperiment,
    SingleCellExperiment,
    package = "AcidTest"
)

## SummarizedExperiment ====
object <- RangedSummarizedExperiment
plotCountsPerBroadClass(object)

## SingleCellExperiment ====
object <- SingleCellExperiment
plotCountsPerBroadClass(object)
}
\author{
Michael Steinbaugh, Rory Kirchner
}
