% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotCountsPerCell-methods.R
\name{plotCountsPerCell}
\alias{plotCountsPerCell}
\alias{plotCountsPerCell,SingleCellExperiment-method}
\title{Plot counts per cell}
\usage{
plotCountsPerCell(object, ...)

\S4method{plotCountsPerCell}{SingleCellExperiment}(
  object,
  geom = c("histogram", "ecdf", "violin", "ridgeline", "boxplot"),
  interestingGroups = NULL,
  min = 0L,
  max = Inf,
  point = c("none", "inflection", "knee"),
  trans = "log10",
  title = "Counts per cell"
)
}
\arguments{
\item{object}{Object.}

\item{geom}{\code{character(1)}.
Plot type. Uses \code{\link[base:match.arg]{match.arg()}} internally and defaults to
the first argument in the \code{character} vector.}

\item{interestingGroups}{\code{character}.
Groups of interest to use for visualization.
Corresponds to factors describing the columns of the object.}

\item{min}{\code{numeric(1)}.
Recommended minimum value cutoff.}

\item{max}{\code{numeric(1)}.
Recommended maximum value cutoff.}

\item{point}{\code{character(1)}.
Label either the knee or inflection points per sample.
Requires \code{geom = "ecdf"}.}

\item{trans}{\code{character(1)}.
Name of the axis scale transformation to apply.

For more information:

\if{html}{\out{<div class="sourceCode">}}\preformatted{help(topic = "scale_x_continuous", package = "ggplot2")
}\if{html}{\out{</div>}}}

\item{title}{\code{character(1)}.
Title.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Plot the disambiguated counts per cell.
}
\details{
"Counts" refer to universal molecular identifier (UMI) counts for
droplet-based scRNA-seq data.
}
\note{
Updated 2022-03-07.
}
\examples{
data(SingleCellExperiment_splatter, package = "AcidTest")

## SingleCellExperiment ====
object <- SingleCellExperiment_splatter
plotCountsPerCell(object, geom = "violin")
plotCountsPerCell(object, geom = "ridgeline")
plotCountsPerCell(object, geom = "ecdf")
plotCountsPerCell(object, geom = "histogram")
plotCountsPerCell(object, geom = "boxplot")
}
\seealso{
\code{\link[AcidGenerics:plotCountsVsFeatures]{plotCountsVsFeatures()}}.
}
\author{
Michael Steinbaugh, Rory Kirchner
}
