% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCountsPerFeature-methods.R
\docType{methods}
\name{plotCountsPerFeature}
\alias{plotCountsPerFeature}
\alias{plotCountsPerFeature,SummarizedExperiment-method}
\alias{plotCountsPerFeature,SingleCellExperiment-method}
\title{Plot counts per feature}
\usage{
plotCountsPerFeature(object, ...)

\S4method{plotCountsPerFeature}{SummarizedExperiment}(object, assay = 1L,
  interestingGroups = NULL, geom = c("boxplot", "density", "jitter"),
  trans = c("identity", "log2", "log10"), color = getOption(x =
  "acid.color.discrete", default = acidplots::scale_color_synesthesia_d()),
  fill = getOption(x = "acid.fill.discrete", default =
  acidplots::scale_fill_synesthesia_d()), labels = list(title =
  "Counts per feature", subtitle = NULL, samplesAxis = NULL, countsAxis =
  "counts"), flip = getOption(x = "acid.flip", default = TRUE),
  minMethod = c("absolute", "perRow"))

\S4method{plotCountsPerFeature}{SingleCellExperiment}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{assay}{\code{vector(1)}.
Assay name or index position.}

\item{interestingGroups}{\code{character}.
Groups of interest to use for visualization.
Corresponds to factors describing the columns of the object.}

\item{geom}{\code{character(1)}.
Type of ggplot2 geometric object to use.}

\item{trans}{\code{character(1)}.
Apply a log transformation (e.g. \code{log2(x + 1L)}) to the count matrix prior
to melting, if desired. Use \code{"identity"} to return unmodified (default).}

\item{color}{\code{ScaleDiscrete}.
Desired ggplot2 color scale. Must supply discrete values. When set \code{NULL},
the default ggplot2 color palette will be used. If manual color definitions
are desired, we recommend using \code{\link[ggplot2:scale_color_manual]{ggplot2::scale_color_manual()}}.

To set the discrete color palette globally, use:\preformatted{options(acid.color.discrete = ggplot2::scale_color_viridis_d())
}}

\item{fill}{\code{ggproto}/\code{ScaleDiscrete}.
Desired ggplot2 fill scale. Must supply discrete values. When set to
\code{NULL}, the default ggplot2 color palette will be used. If manual color
definitions are desired, we recommend using \code{\link[ggplot2:scale_fill_manual]{ggplot2::scale_fill_manual()}}.

To set the discrete fill palette globally, use:\preformatted{options(acid.fill.discrete = ggplot2::scale_fill_viridis_d())
}}

\item{labels}{\code{list}.
ggplot2 labels.
See \code{\link[ggplot2:labs]{ggplot2::labs()}} for details.}

\item{flip}{\code{logical(1)}.
Flip x and y axes. Recommended for plots containing many samples.}

\item{minMethod}{\code{character(1)}.
Only applies when \code{min} argument is numeric.
Uses \code{\link[base:match.arg]{match.arg()}}.
\itemize{
\item \code{absolute}: Applies hard cutoff to \code{counts} column after the melt
operation. This applies to all counts, not per feature.
\item \code{perRow}: Applies cutoff per row (i.e. gene). Internally,
\code{\link[base:rowSums]{rowSums()}} values are checked against this cutoff
threshold prior to the melt operation.
}}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Generally, we expect similar count spreads for all genes between samples
unless the library sizes or total RNA expression are different.
}
\section{Methods (by class)}{
\itemize{
\item \code{SingleCellExperiment}: Applies \code{\link[=aggregateCellsToSamples]{aggregateCellsToSamples()}}
calculation to summarize at sample level prior to plotting.\cr
Passes \code{...} to \code{SummarizedExperiment} method.
}}

\note{
Updated 2019-09-15.
}
\examples{
data(
    RangedSummarizedExperiment,
    SingleCellExperiment,
    package = "acidtest"
)

## SummarizedExperiment ====
object <- RangedSummarizedExperiment
plotCountsPerFeature(object, geom = "boxplot")
plotCountsPerFeature(object, geom = "density")

## SingleCellExperiment ====
object <- SingleCellExperiment
plotCountsPerFeature(object)
}
