% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCountsPerFeature-methods.R
\docType{methods}
\name{plotCountsPerFeature}
\alias{plotCountsPerFeature}
\alias{plotCountsPerFeature,SummarizedExperiment-method}
\alias{plotCountsPerFeature,SingleCellExperiment-method}
\title{Plot counts per feature}
\usage{
plotCountsPerFeature(object, ...)

\S4method{plotCountsPerFeature}{SummarizedExperiment}(object, assay = 1L,
  minCounts = 1L, minCountsMethod = c("perFeature", "absolute"),
  interestingGroups = NULL, geom = c("boxplot", "density"),
  trans = c("identity", "log2", "log10"), color = getOption(x =
  "acid.color.discrete", default =
  acidplots::scale_colour_synesthesia_d()), fill = getOption(x =
  "acid.fill.discrete", default = acidplots::scale_fill_synesthesia_d()),
  flip = getOption(x = "acid.flip", default = TRUE),
  countsAxisLabel = "counts", title = "Counts per feature")

\S4method{plotCountsPerFeature}{SingleCellExperiment}(object, assay = 1L,
  minCounts = 1L, minCountsMethod = c("perFeature", "absolute"),
  interestingGroups = NULL, geom = c("boxplot", "density"),
  trans = c("identity", "log2", "log10"), color = getOption(x =
  "acid.color.discrete", default =
  acidplots::scale_colour_synesthesia_d()), fill = getOption(x =
  "acid.fill.discrete", default = acidplots::scale_fill_synesthesia_d()),
  flip = getOption(x = "acid.flip", default = TRUE),
  countsAxisLabel = "counts", title = "Counts per feature")
}
\arguments{
\item{object}{Object.}

\item{assay}{\code{vector(1)}.
Name or index of count matrix slotted in
\code{\link[SummarizedExperiment:assays]{assays()}}. When passing in a string, the
name must be defined in
\code{\link[SummarizedExperiment:assayNames]{assayNames()}}.}

\item{minCounts}{\code{integer(1)}.
Minimum count threshold to apply. Filters using "greater than or equal to"
logic internally. Note that this threshold gets applied prior to
logarithmic transformation, when \code{trans} argument applies.}

\item{minCountsMethod}{\code{character(1)}.
Uses \code{\link[base:match.arg]{match.arg()}}.
\itemize{
\item \code{perFeature}: \emph{Recommended}. Applies cutoff per row feature (i.e. gene).
Internally, \code{\link[base:rowSums]{rowSums()}} values are checked against this
cutoff threshold prior to the melt operation.
\item \code{absolute}: Applies hard cutoff to \code{counts} column after the melt
operation. This applies to all counts, not per feature.
}}

\item{interestingGroups}{\code{character}.
Groups of interest that define the samples. If left unset, defaults to
\code{sampleName}.}

\item{geom}{\code{character(1)}.
Type of ggplot2 geometric object to use.}

\item{trans}{\code{character(1)}.
Apply a log transformation (e.g. \code{log2(x + 1L)}) to the count matrix prior
to melting, if desired. Use \code{"identity"} to return unmodified (default).}

\item{color}{\code{ScaleDiscrete}.
Desired ggplot2 color scale. Must supply discrete values. When set \code{NULL},
the default ggplot2 color palette will be used. If manual color definitions
are desired, we recommend using \code{\link[ggplot2:scale_color_manual]{ggplot2::scale_color_manual()}}.

To set the discrete color palette globally, use:\preformatted{options(acid.color.discrete = ggplot2::scale_color_viridis_d())
}}

\item{fill}{\code{ggproto}/\code{ScaleDiscrete}.
Desired ggplot2 fill scale. Must supply discrete values. When set to
\code{NULL}, the default ggplot2 color palette will be used. If manual color
definitions are desired, we recommend using \code{\link[ggplot2:scale_fill_manual]{ggplot2::scale_fill_manual()}}.

To set the discrete fill palette globally, use:\preformatted{options(acid.fill.discrete = ggplot2::scale_fill_viridis_d())
}}

\item{flip}{\code{logical(1)}.
Flip x and y axes. Recommended for plots containing many samples.}

\item{countsAxisLabel}{\code{character(1)}.
Counts axis label.}

\item{title}{\code{character(1)}.
Plot title.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Generally, we expect similar count spreads for all genes between samples
unless the library sizes or total RNA expression are different.
}
\examples{
data(rse, sce, package = "acidtest")

## SummarizedExperiment ====
plotCountsPerFeature(rse, geom = "boxplot")
plotCountsPerFeature(rse, geom = "density")

## SingleCellExperiment ====
plotCountsPerFeature(sce)
}
