% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotFeaturesDetected-methods.R
\name{plotFeaturesDetected}
\alias{plotFeaturesDetected}
\alias{plotFeaturesDetected,SingleCellExperiment-method}
\alias{plotFeaturesDetected,SummarizedExperiment-method}
\title{Plot features detected}
\usage{
plotFeaturesDetected

\S4method{plotFeaturesDetected}{SingleCellExperiment}(object, ...)

\S4method{plotFeaturesDetected}{SummarizedExperiment}(
  object,
  assay = 1L,
  interestingGroups = NULL,
  limit = 0L,
  minCounts = 1L,
  labels = list(title = "Features detected", subtitle = NULL, x = NULL, y = "features"),
  flip = getOption(x = "acid.flip", default = TRUE)
)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{assay}{\code{vector(1)}.
Assay name or index position.}

\item{interestingGroups}{\code{character}.
Groups of interest to use for visualization.
Corresponds to factors describing the columns of the object.}

\item{limit}{\code{numeric(1)}.
Threshold limit.}

\item{minCounts}{\code{integer(1)}.
Minimum number of counts per feature (i.e. gene).}

\item{labels}{\code{list}.
ggplot2 labels.
See \code{\link[ggplot2:labs]{ggplot2::labs()}} for details.}

\item{flip}{\code{logical(1)}.
Flip x and y axes. Recommended for plots containing many samples.}
}
\value{
\code{ggplot}.
}
\description{
Visualize the number of features (i.e. genes) detected.
}
\section{Functions}{
\itemize{
\item \code{plotFeaturesDetected,SingleCellExperiment-method}: Applies \code{aggregateCellsToSamples()}
calculation to summarize at sample level prior to plotting.\cr
Passes \code{...} to \code{SummarizedExperiment} method.
}}

\note{
Updated 2022-03-04.
}
\examples{
data(
    RangedSummarizedExperiment,
    SingleCellExperiment_splatter,
    package = "AcidTest"
)

## SummarizedExperiment ====
object <- RangedSummarizedExperiment
plotFeaturesDetected(object)

## SingleCellExperiment ====
object <- SingleCellExperiment_splatter
plotFeaturesDetected(object)
}
