% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFeaturesDetected-methods.R
\docType{methods}
\name{plotFeaturesDetected}
\alias{plotFeaturesDetected}
\alias{plotFeaturesDetected,SummarizedExperiment-method}
\alias{plotFeaturesDetected,SingleCellExperiment-method}
\title{Plot features detected}
\usage{
plotFeaturesDetected

\S4method{plotFeaturesDetected}{SummarizedExperiment}(object, assay = 1L,
  interestingGroups = NULL, limit = 0L, minCounts = 1L,
  fill = getOption(x = "acid.fill.discrete", default =
  acidplots::scale_fill_synesthesia_d()), flip = getOption(x =
  "acid.flip", default = TRUE), labels = list(title =
  "Features detected", subtitle = NULL, x = NULL, y = "features"))

\S4method{plotFeaturesDetected}{SingleCellExperiment}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{assay}{\code{vector(1)}.
Assay name or index position.}

\item{interestingGroups}{\code{character}.
Groups of interest to use for visualization.
Corresponds to factors describing the columns of the object.}

\item{limit}{\code{numeric(1)}.
Threshold limit.}

\item{minCounts}{\code{integer(1)}.
Minimum number of counts per feature (i.e. gene).}

\item{fill}{\code{ggproto}/\code{ScaleDiscrete}.
Desired ggplot2 fill scale. Must supply discrete values. When set to
\code{NULL}, the default ggplot2 color palette will be used. If manual color
definitions are desired, we recommend using \code{\link[ggplot2:scale_fill_manual]{ggplot2::scale_fill_manual()}}.

To set the discrete fill palette globally, use:\preformatted{options(acid.fill.discrete = ggplot2::scale_fill_viridis_d())
}}

\item{flip}{\code{logical(1)}.
Flip x and y axes. Recommended for plots containing many samples.}

\item{labels}{\code{list}.
ggplot2 labels.
See \code{\link[ggplot2:labs]{ggplot2::labs()}} for details.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Visualize the number of features (i.e. genes) detected.
}
\section{Methods (by class)}{
\itemize{
\item \code{SingleCellExperiment}: Applies \code{\link[=aggregateCellsToSamples]{aggregateCellsToSamples()}}
calculation to summarize at sample level prior to plotting.\cr
Passes \code{...} to \code{SummarizedExperiment} method.
}}

\note{
Updated 2019-09-15.
}
\examples{
data(
    RangedSummarizedExperiment,
    SingleCellExperiment,
    package = "acidtest"
)

## SummarizedExperiment ====
object <- RangedSummarizedExperiment
plotFeaturesDetected(object)

## SingleCellExperiment ====
object <- SingleCellExperiment
plotFeaturesDetected(object)
}
