% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFeaturesPerCell-methods.R
\name{plotFeaturesPerCell}
\alias{plotFeaturesPerCell}
\alias{plotFeaturesPerCell,SingleCellExperiment-method}
\title{Plot features per cell}
\usage{
plotFeaturesPerCell(object, ...)

\S4method{plotFeaturesPerCell}{SingleCellExperiment}(
  object,
  geom = c("histogram", "ecdf", "violin", "ridgeline", "boxplot"),
  interestingGroups = NULL,
  min = 0L,
  max = Inf,
  trans = "log2",
  fill = getOption(x = "acid.fill.discrete", default =
    acidplots::scale_fill_synesthesia_d()),
  labels = list(title = "Features per cell", subtitle = NULL, metricAxis = "features",
    otherAxis = NULL)
)
}
\arguments{
\item{object}{Object.}

\item{geom}{\code{character(1)}.
Plot type. Uses \code{\link[base:match.arg]{match.arg()}} internally and defaults to
the first argument in the \code{character} vector.}

\item{interestingGroups}{\code{character}.
Groups of interest to use for visualization.
Corresponds to factors describing the columns of the object.}

\item{min}{\code{numeric(1)}.
Recommended minimum value cutoff.}

\item{max}{\code{numeric(1)}.
Recommended maximum value cutoff.}

\item{trans}{\code{character(1)}.
Name of the axis scale transformation to apply.

For more information:\preformatted{help(topic = "scale_x_continuous", package = "ggplot2")
}}

\item{fill}{\code{ggproto}/\code{ScaleDiscrete}.
Desired ggplot2 fill scale. Must supply discrete values. When set to
\code{NULL}, the default ggplot2 color palette will be used. If manual color
definitions are desired, we recommend using \code{\link[ggplot2:scale_fill_manual]{ggplot2::scale_fill_manual()}}.

To set the discrete fill palette globally, use:\preformatted{options(acid.fill.discrete = ggplot2::scale_fill_viridis_d())
}}

\item{labels}{\code{list}.
ggplot2 labels.
See \code{\link[ggplot2:labs]{ggplot2::labs()}} for details.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Plot features per cell
}
\note{
Updated 2019-09-16.
}
\examples{
data(SingleCellExperiment, package = "acidtest")

## SingleCellExperiment ====
object <- SingleCellExperiment
object <- calculateMetrics(object)
plotFeaturesPerCell(object)
}
\author{
Michael Steinbaugh, Rory Kirchner
}
