% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGenderMarkers-methods.R
\name{plotGenderMarkers}
\alias{plotGenderMarkers}
\alias{plotGenderMarkers,SummarizedExperiment-method}
\alias{plotGenderMarkers,SingleCellExperiment-method}
\title{Plot sexually dimorphic gender marker genes}
\usage{
plotGenderMarkers(object, ...)

\S4method{plotGenderMarkers}{SummarizedExperiment}(object, style = "wide", ...)

\S4method{plotGenderMarkers}{SingleCellExperiment}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{style}{\code{character(1)}.
Plot style.}

\item{...}{Passthrough to \code{\link[=plotCounts]{plotCounts()}}, with the \code{genes} argument
automatically defined.}
}
\value{
\code{ggplot}.
}
\description{
This is a convenience function that quickly plots known sexually dimorphic
genes, and overall gene experssion for the X and Y chromosomes.
}
\section{Functions}{
\itemize{
\item \code{plotGenderMarkers,SingleCellExperiment-method}: Applies \code{\link[=pseudobulk]{pseudobulk()}} calculation to average
gene expression at sample level prior to plotting.\cr
Passes \code{...} to \code{SummarizedExperiment} method.
}}

\note{
Currently only \emph{Homo sapiens} and \emph{Mus musculus} genomes are supported.

Updated 2019-09-15.
}
\examples{
data(
    RangedSummarizedExperiment,
    SingleCellExperiment,
    package = "acidtest"
)

## SummarizedExperiment ====
object <- RangedSummarizedExperiment
plotGenderMarkers(object)

## SingleCellExperiment ====
object <- SingleCellExperiment
plotGenderMarkers(object)
}
\seealso{
\code{plotCounts()}.
}
