% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotMitoRatio-methods.R
\name{plotMitoRatio}
\alias{plotMitoRatio}
\alias{plotMitoRatio,SingleCellExperiment-method}
\title{Plot mitochondrial transcript abundance}
\usage{
plotMitoRatio(object, ...)

\S4method{plotMitoRatio}{SingleCellExperiment}(
  object,
  geom = c("histogram", "ecdf", "violin", "ridgeline", "boxplot"),
  interestingGroups = NULL,
  max = 1L,
  trans = "sqrt",
  labels = list(title = "Mito ratio", subtitle = NULL, metricAxis = NULL, otherAxis =
    NULL)
)
}
\arguments{
\item{object}{Object.}

\item{geom}{\code{character(1)}.
Plot type. Uses \code{\link[base:match.arg]{match.arg()}} internally and defaults to
the first argument in the \code{character} vector.}

\item{interestingGroups}{\code{character}.
Groups of interest to use for visualization.
Corresponds to factors describing the columns of the object.}

\item{max}{\code{numeric(1)}.
Recommended maximum value cutoff.}

\item{trans}{\code{character(1)}.
Name of the axis scale transformation to apply.

For more information:\preformatted{help(topic = "scale_x_continuous", package = "ggplot2")
}}

\item{labels}{\code{list}.
ggplot2 labels.
See \code{\link[ggplot2:labs]{ggplot2::labs()}} for details.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Plot mitochondrial transcript abundance
}
\note{
Updated 2022-01-19.
}
\examples{
data(SingleCellExperiment_splatter, package = "AcidTest")

## SingleCellExperiment ====
object <- SingleCellExperiment_splatter
try({
    plotMitoRatio(object)
})
}
\author{
Michael Steinbaugh, Rory Kirchner
}
