% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMitoVsCoding-methods.R
\name{plotMitoVsCoding}
\alias{plotMitoVsCoding}
\alias{plotMitoVsCoding,SingleCellExperiment-method}
\title{Plot mitochondrial vs. coding counts}
\usage{
plotMitoVsCoding(object, ...)

\S4method{plotMitoVsCoding}{SingleCellExperiment}(
  object,
  interestingGroups = NULL,
  trendline = FALSE,
  color = getOption(x = "acid.color.discrete", default =
    acidplots::scale_color_synesthesia_d()),
  trans = "log2",
  labels = list(title = "Mito vs. coding", subtitle = NULL, x = "coding", y = "mito")
)
}
\arguments{
\item{object}{Object.}

\item{interestingGroups}{\code{character}.
Groups of interest to use for visualization.
Corresponds to factors describing the columns of the object.}

\item{trendline}{\code{logical(1)}.
Include trendline on plot.}

\item{color}{\code{ScaleDiscrete}.
Desired ggplot2 color scale. Must supply discrete values. When set \code{NULL},
the default ggplot2 color palette will be used. If manual color definitions
are desired, we recommend using \code{\link[ggplot2:scale_manual]{ggplot2::scale_color_manual()}}.

To set the discrete color palette globally, use:\preformatted{options(acid.color.discrete = ggplot2::scale_color_viridis_d())
}}

\item{trans}{\code{character(1)}.
Name of the axis scale transformation to apply.

For more information:\preformatted{help(topic = "scale_x_continuous", package = "ggplot2")
}}

\item{labels}{\code{list}.
ggplot2 labels.
See \code{\link[ggplot2:labs]{ggplot2::labs()}} for details.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Plot mitochondrial vs. coding counts
}
\note{
Updated 2019-09-16.
}
\examples{
data(SingleCellExperiment, package = "acidtest")

## SingleCellExperiment ====
object <- SingleCellExperiment
object <- calculateMetrics(object)
if (!anyNA(object$nMito)) {
    plotMitoVsCoding(object)
}
}
\author{
Michael Steinbaugh, Rory Kirchner
}
