% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQC-methods.R
\docType{methods}
\name{plotQC}
\alias{plotQC}
\alias{plotQC,SummarizedExperiment-method}
\title{Quality control}
\usage{
plotQC(object, ...)

\S4method{plotQC}{SummarizedExperiment}(object, assay = 1L)
}
\arguments{
\item{object}{Object.}

\item{assay}{\code{vector(1)}.
Assay name or index position.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Quality control
}
\note{
Updated 2019-07-29.
}
\examples{
data(RangedSummarizedExperiment, package = "acidtest")
rse <- RangedSummarizedExperiment

## SummarizedExperiment ====
plotQC(rse)
}
