% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQC-methods.R
\docType{methods}
\name{plotQC}
\alias{plotQC}
\alias{plotQC,SummarizedExperiment-method}
\title{Quality control}
\usage{
plotQC(object, ...)

\S4method{plotQC}{SummarizedExperiment}(object, assay = 1L)
}
\arguments{
\item{object}{Object.}

\item{assay}{\code{vector(1)}.
Name or index of count matrix slotted in
\code{\link[SummarizedExperiment:assays]{assays()}}. When passing in a string, the
name must be defined in
\code{\link[SummarizedExperiment:assayNames]{assayNames()}}.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Quality control
}
\examples{
data(RangedSummarizedExperiment, package = "acidtest")
rse <- RangedSummarizedExperiment

## SummarizedExperiment ====
plotQC(rse)
}
