% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotQc-methods.R
\name{plotQc}
\alias{plotQc}
\alias{plotQc,SingleCellExperiment-method}
\alias{plotQc,SummarizedExperiment-method}
\title{Quality control}
\usage{
plotQc(object, ...)

\S4method{plotQc}{SingleCellExperiment}(
  object,
  assay = 1L,
  interestingGroups = NULL,
  geom = c("histogram", "ecdf", "violin", "ridgeline", "boxplot"),
  legend = getOption(x = "acid.legend", default = TRUE),
  labels = list(title = "Quality control", subtitle = NULL)
)

\S4method{plotQc}{SummarizedExperiment}(
  object,
  assay = 1L,
  interestingGroups = NULL,
  legend = getOption(x = "acid.legend", default = TRUE),
  labels = list(title = "Quality control", subtitle = NULL)
)
}
\arguments{
\item{object}{Object.}

\item{assay}{\code{vector(1)}.
Assay name or index position.}

\item{interestingGroups}{\code{character}.
Groups of interest to use for visualization.
Corresponds to factors describing the columns of the object.}

\item{geom}{\code{character(1)}.
Plot type. Uses \code{\link[base:match.arg]{match.arg()}} internally and defaults to
the first argument in the \code{character} vector.}

\item{legend}{\code{logical(1)}.
Include plot legend.}

\item{labels}{\code{list}.
ggplot2 labels.
See \code{ggplot2::labs()} for details.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Quality control
}
\note{
Updated 2023-08-11.
}
\examples{
data(
    RangedSummarizedExperiment,
    SingleCellExperiment_splatter,
    package = "AcidTest"
)

## SummarizedExperiment ====
object <- RangedSummarizedExperiment
plotQc(object)

## SingleCellExperiment ====
object <- SingleCellExperiment_splatter
plotQc(object)
}
