% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotQC-methods.R
\name{plotQC}
\alias{plotQC}
\alias{plotQC,SingleCellExperiment-method}
\alias{plotQC,SummarizedExperiment-method}
\title{Quality control}
\usage{
plotQC(object, ...)

\S4method{plotQC}{SingleCellExperiment}(
  object,
  assay = 1L,
  geom = c("histogram", "ecdf", "violin", "ridgeline", "boxplot"),
  legend = getOption(x = "acid.legend", default = TRUE),
  labels = list(title = "Quality control", subtitle = NULL)
)

\S4method{plotQC}{SummarizedExperiment}(
  object,
  assay = 1L,
  interestingGroups = NULL,
  legend = getOption(x = "acid.legend", default = TRUE),
  labels = list(title = "Quality control", subtitle = NULL)
)
}
\arguments{
\item{object}{Object.}

\item{assay}{\code{vector(1)}.
Assay name or index position.}

\item{geom}{\code{character(1)}.
Plot type. Uses \code{\link[base:match.arg]{match.arg()}} internally and defaults to
the first argument in the \code{character} vector.}

\item{legend}{\code{logical(1)}.
Include plot legend.}

\item{labels}{\code{list}.
ggplot2 labels.
See \code{\link[ggplot2:labs]{ggplot2::labs()}} for details.}

\item{interestingGroups}{\code{character}.
Groups of interest to use for visualization.
Corresponds to factors describing the columns of the object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Quality control
}
\note{
Updated 2021-09-10.
}
\examples{
data(
    RangedSummarizedExperiment,
    SingleCellExperiment,
    package = "AcidTest"
)

## SummarizedExperiment ====
object <- RangedSummarizedExperiment
plotQC(object)

## SingleCellExperiment ====
object <- SingleCellExperiment
object <- calculateMetrics(object)
plotQC(object)
}
