% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQuantileHeatmap-methods.R
\docType{methods}
\name{plotQuantileHeatmap}
\alias{plotQuantileHeatmap}
\alias{plotQuantileHeatmap,SummarizedExperiment-method}
\alias{plotQuantileHeatmap,SingleCellExperiment-method}
\title{Quantile heatmap}
\usage{
\S4method{plotQuantileHeatmap}{SummarizedExperiment}(object, assay = 1L,
  interestingGroups = NULL, n = 10L, clusterRows = TRUE,
  clusterCols = TRUE, showRownames = FALSE, showColnames = TRUE,
  treeheightRow = 0L, treeheightCol = 50L, color = getOption(x =
  "acid.heatmap.color", default = acidplots::synesthesia),
  legendColor = getOption(x = "acid.heatmap.color", default =
  acidplots::synesthesia), legend = FALSE, borderColor = NULL,
  title = NULL, ...)

\S4method{plotQuantileHeatmap}{SingleCellExperiment}(object, assay = 1L,
  interestingGroups = NULL, n = 10L, clusterRows = TRUE,
  clusterCols = TRUE, showRownames = FALSE, showColnames = TRUE,
  treeheightRow = 0L, treeheightCol = 50L, color = getOption(x =
  "acid.heatmap.color", default = acidplots::synesthesia),
  legendColor = getOption(x = "acid.heatmap.color", default =
  acidplots::synesthesia), legend = FALSE, borderColor = NULL,
  title = NULL, ...)
}
\arguments{
\item{object}{Object.}

\item{assay}{\code{vector(1)}.
Name or index of count matrix slotted in
\code{\link[SummarizedExperiment:assays]{assays()}}. When passing in a string, the
name must be defined in
\code{\link[SummarizedExperiment:assayNames]{assayNames()}}.}

\item{interestingGroups}{\code{character}.
Groups of interest that define the samples. If left unset, defaults to
\code{sampleName}.}

\item{n}{\code{integer(1)}.
The number of quantile breaks to create.}

\item{clusterRows}{\code{logical(1)}.
Arrange with hierarchical clustering.}

\item{clusterCols}{\code{logical(1)}.
Arrange with hierarchical clustering.}

\item{showRownames}{\code{logical(1)}.
Show row or column names.}

\item{showColnames}{\code{logical(1)}.
Show row or column names.}

\item{treeheightRow}{\code{integer(1)}.
Size of the row and column dendrograms. Use \code{0} to disable.}

\item{treeheightCol}{\code{integer(1)}.
Size of the row and column dendrograms. Use \code{0} to disable.}

\item{color}{\code{ScaleDiscrete}.
Desired ggplot2 color scale. Must supply discrete values. When set \code{NULL},
the default ggplot2 color palette will be used. If manual color definitions
are desired, we recommend using \code{\link[ggplot2:scale_color_manual]{ggplot2::scale_color_manual()}}.

To set the discrete color palette globally, use:\preformatted{options(acid.color.discrete = ggplot2::scale_color_viridis_d())
}}

\item{legendColor}{\code{function} or \code{NULL}.
Hexadecimal color function to use for legend labels. Note that hexadecimal
values are not supported. If set \code{NULL}, will use the default pheatmap
colors.}

\item{legend}{\code{logical(1)}.
Show the color legend.}

\item{borderColor}{\code{character(1)} or \code{NULL}.
Border color. Disabled by default for improved aesthetics.}

\item{title}{\code{character(1)}.
Plot title.}

\item{...}{Additional arguments.}
}
\value{
\code{pheatmap}.
}
\description{
Quantile heatmap
}
\examples{
data(rse, sce, package = "acidtest")

## SummarizedExperiment ====
plotQuantileHeatmap(rse)

## SingleCellExperiment ====
plotQuantileHeatmap(sce)
}
