% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotSums-methods.R
\name{plotSums}
\alias{plotSums}
\alias{plotSums,matrix-method}
\alias{plotSums,Matrix-method}
\alias{plotSums,SummarizedExperiment-method}
\title{Plot row or column sums}
\usage{
plotSums(object, ...)

\S4method{plotSums}{matrix}(object, MARGIN = 1L)

\S4method{plotSums}{Matrix}(object, MARGIN = 1L)

\S4method{plotSums}{SummarizedExperiment}(object, assay = 1L, MARGIN = 1L)
}
\arguments{
\item{object}{Object.}

\item{MARGIN}{\code{integer(1-2)}.
Dimension where the function will be applied.
For a two-dimensional matrix:
\code{1} indicates rows;
\code{2} indicates columns;
\code{c(1, 2)} indicates rows and columns.}

\item{assay}{\code{vector(1)}.
Assay name or index position.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Visualize row or column sums using empirical cumulative distribution function
(ECDF) plot.
}
\note{
Updated 2019-09-15.
}
\examples{
data(
    RangedSummarizedExperiment,
    SingleCellExperiment,
    package = "AcidTest"
)

## matrix ====
object <- basejump::counts(RangedSummarizedExperiment)
plotSums(object, MARGIN = 1L)

## Matrix ====
object <- basejump::counts(SingleCellExperiment)
plotSums(object, MARGIN = 2L)

## SummarizedExperiment ====
object <- SingleCellExperiment
plotSums(object)

## SingleCellExperiment
object <- RangedSummarizedExperiment
plotSums(object)
}
