% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotSums-methods.R
\name{plotSums}
\alias{plotSums}
\alias{plotSums,SummarizedExperiment-method}
\title{Plot row or column sums}
\usage{
plotSums(object, ...)

\S4method{plotSums}{SummarizedExperiment}(
  object,
  MARGIN,
  assay = 1L,
  interestingGroups = NULL,
  labels = list(title = NULL, subtitle = NULL)
)
}
\arguments{
\item{object}{Object.}

\item{MARGIN}{\code{integer(1-2)}.
Dimension where the function will be applied.
For a two-dimensional matrix:
\code{1} indicates rows;
\code{2} indicates columns;
\code{c(1, 2)} indicates rows and columns.}

\item{assay}{\code{vector(1)}.
Assay name or index position.}

\item{interestingGroups}{\code{character}.
Groups of interest to use for visualization.
Corresponds to factors describing the columns of the object.}

\item{labels}{\code{list}.
ggplot2 labels.
See \code{ggplot2::labs()} for details.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Visualize row or column sums using empirical cumulative distribution function
(ECDF) plot.
}
\note{
Updated 2021-09-10.
}
\examples{
data(
    RangedSummarizedExperiment,
    package = "AcidTest"
)

## SummarizedExperiment ====
object <- RangedSummarizedExperiment
plotSums(object, MARGIN = 1L)
plotSums(object, MARGIN = 2L)
}
