% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotTotalCounts-methods.R
\name{plotTotalCounts}
\alias{plotTotalCounts}
\alias{plotTotalCounts,SingleCellExperiment-method}
\alias{plotTotalCounts,SummarizedExperiment-method}
\title{Plot total read counts}
\usage{
plotTotalCounts(object, ...)

\S4method{plotTotalCounts}{SingleCellExperiment}(object, ...)

\S4method{plotTotalCounts}{SummarizedExperiment}(
  object,
  assay = 1L,
  interestingGroups = NULL,
  perMillion = FALSE,
  fill = getOption(x = "acid.fill.discrete", default = getOption(x =
    "ggplot2.discrete.fill", default = AcidPlots::scale_fill_synesthesia_d())),
  labels = list(title = "Total counts", subtitle = NULL, x = NULL, y = "counts"),
  flip = getOption(x = "acid.flip", default = TRUE)
)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{assay}{\code{vector(1)}.
Assay name or index position.}

\item{interestingGroups}{\code{character}.
Groups of interest to use for visualization.
Corresponds to factors describing the columns of the object.}

\item{perMillion}{\code{logical(1)}.
Display as counts per million.}

\item{fill}{\code{ggproto}/\code{ScaleDiscrete}.
Desired ggplot2 fill scale. Must supply discrete values. When set to
\code{NULL}, the default ggplot2 color palette will be used. If manual color
definitions are desired, we recommend using \code{ggplot2::scale_fill_manual()}.

To set the discrete fill palette globally, use:\preformatted{options(acid.fill.discrete = ggplot2::scale_fill_viridis_d())
}}

\item{labels}{\code{list}.
ggplot2 labels.
See \code{\link[ggplot2:labs]{ggplot2::labs()}} for details.}

\item{flip}{\code{logical(1)}.
Flip x and y axes. Recommended for plots containing many samples.}
}
\value{
\code{ggplot}.
}
\description{
Plot total read counts
}
\section{Functions}{
\itemize{
\item \code{plotTotalCounts,SingleCellExperiment-method}: Applies \code{\link[=aggregateCellsToSamples]{aggregateCellsToSamples()}} calculation
to summarize at sample level prior to plotting.\cr
Passes \code{...} to \code{SummarizedExperiment} method.
}}

\note{
Updated 2021-02-16.
}
\examples{
data(
    RangedSummarizedExperiment,
    SingleCellExperiment,
    package = "AcidTest"
)

## SummarizedExperiment ====
object <- RangedSummarizedExperiment
plotTotalCounts(object)

## SingleCellExperiment ====
object <- SingleCellExperiment
plotTotalCounts(object)
}
