% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotUpset-methods.R
\name{plotUpset}
\alias{plotUpset}
\alias{plotUpset,list-method}
\alias{plotUpset,matrix-method}
\alias{plotUpset,data.frame-method}
\alias{plotUpset,DataFrame-method}
\title{UpSet plot}
\usage{
plotUpset(object, ...)

\S4method{plotUpset}{list}(object, ...)

\S4method{plotUpset}{matrix}(
  object,
  nIntersections = 40L,
  orderBySize = c(intersections = TRUE, sets = TRUE),
  labels = list(title = NULL, subtitle = NULL)
)

\S4method{plotUpset}{data.frame}(object, ...)

\S4method{plotUpset}{DataFrame}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{nIntersections}{\code{integer(1)} or \code{Inf}.
Maximum number of intersections to plot.
Set \code{Inf} to plot all intersections.}

\item{orderBySize}{\code{logical}.
Whether to order main bar plot and/or intersection matrix by set size.
\itemize{
\item "bars" refers to main bar plot.
\item "matrix" refers to intersection matrix, shown as connected dots.
}

Can pass in \code{TRUE}/\code{FALSE} boolean flag and both "bars" and "matrix"
settings will inherit.}

\item{labels}{\code{list}.
ggplot2 labels.
See \code{\link[ggplot2:labs]{ggplot2::labs()}} for details.}

\item{sortIntersections}{\code{logical(1)}.}

\item{sortSets}{\code{logical(1)}.}
}
\value{
Plot.
}
\description{
UpSet plot
}
\details{
S4 wrapper for \code{UpSetR::upset()} with improved default aesthetics.
}
\note{
Updated 2021-08-12.
}
\examples{
list <- list(
    "aaa" = c("a", "b", "c", "d", "e", "f"),
    "bbb" = c("b", "c", "d", "e", "f", "g"),
    "ccc" = c("c", "d", "e", "f", "g", "h"),
    "ddd" = c("d", "e", "f", "g", "h", "i")
)
print(list)
mat <- intersectionMatrix(list)
print(mat)

## list ====
plotUpset(list)

## matrix ====
plotUpset(mat)
}
\seealso{
\itemize{
\item \code{ComplexUpset::upset}, \code{ComplexUpset::upset_data}.
\item \code{upsetMatrix()}.
\item ComplexUpset package.
\item UpSetR package.
}
}
