% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUpset-methods.R
\name{plotUpset}
\alias{plotUpset}
\alias{plotUpset,list-method}
\alias{plotUpset,matrix-method}
\alias{plotUpset,data.frame-method}
\alias{plotUpset,DataFrame-method}
\title{UpSet plot}
\usage{
plotUpset(object, ...)

\S4method{plotUpset}{list}(object, ...)

\S4method{plotUpset}{matrix}(
  object,
  nIntersects = 40L,
  orderBySize = c(bars = FALSE, matrix = FALSE)
)

\S4method{plotUpset}{data.frame}(object, ...)

\S4method{plotUpset}{DataFrame}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{nIntersects}{\code{integer(1)} or \code{Inf}.
Maximum number of intersections to plot.
Set \code{Inf} to plot all intersections.}

\item{orderBySize}{\code{logical}.
Whether to order main bar plot and/or intersection matrix by set size.
\itemize{
\item "bars" refers to main bar plot.
\item "matrix" refers to intersection matrix, shown as connected dots.
}

Can pass in \code{TRUE}/\code{FALSE} boolean flag and both "bars" and "matrix"
settings will inherit.}
}
\value{
Graphical output, no return.
}
\description{
S4 wrapper for \code{\link[UpSetR:upset]{UpSetR::upset()}} with improved default aesthetics.
}
\note{
Updated 2020-08-25.
}
\examples{
list <- list(
    a = c("a", "b", "c", "d", "e", "f"),
    b = c("b", "c", "d", "e", "f", "g"),
    c = c("c", "d", "e", "f", "g", "h"),
    d = c("d", "e", "f", "g", "h", "i")
)
plotUpset(list)
}
