% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUpset-methods.R
\name{plotUpset}
\alias{plotUpset}
\alias{plotUpset,list-method}
\alias{plotUpset,matrix-method}
\alias{plotUpset,data.frame-method}
\alias{plotUpset,DataFrame-method}
\title{UpSet plot}
\usage{
plotUpset(object, ...)

\S4method{plotUpset}{list}(object, ...)

\S4method{plotUpset}{matrix}(object)

\S4method{plotUpset}{data.frame}(object)

\S4method{plotUpset}{DataFrame}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Graphical output, no return.
}
\description{
S4 wrapper for \code{\link[UpSetR:upset]{UpSetR::upset()}} with improved default aesthetics.
}
\note{
Updated 2020-07-23.
}
\examples{
list <- list(
    a = c("a", "b", "c", "d", "e", "f"),
    b = c("b", "c", "d", "e", "f", "g"),
    c = c("c", "d", "e", "f", "g", "h"),
    d = c("d", "e", "f", "g", "h", "i")
)
plotUpset(list)
}
