% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWaterfall-methods.R
\name{plotWaterfall}
\alias{plotWaterfall}
\alias{plotWaterfall,data.frame-method}
\title{Plot waterfall}
\usage{
plotWaterfall(object, ...)

\S4method{plotWaterfall}{data.frame}(object, xCol, yCol, interestingGroups, labels = NULL)
}
\arguments{
\item{object}{Object.}

\item{interestingGroups}{\code{character}.
Groups of interest to use for visualization.
Corresponds to factors describing the columns of the object.}

\item{labels}{\code{list}.
ggplot2 labels.
See \code{\link[ggplot2:labs]{ggplot2::labs()}} for details.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Plot waterfall
}
\note{
Updated 2019-12-09.
}
\examples{
class(plotWaterfall)
}
