% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWaterfall-methods.R
\name{plotWaterfall}
\alias{plotWaterfall}
\alias{plotWaterfall,data.frame-method}
\title{Plot waterfall}
\usage{
plotWaterfall(object, ...)

\S4method{plotWaterfall}{data.frame}(
  object,
  sampleCol,
  valueCol,
  interestingGroups = NULL,
  label = FALSE,
  fill = getOption(x = "acid.fill.discrete", default =
    acidplots::scale_fill_synesthesia_d()),
  labels = NULL
)
}
\arguments{
\item{object}{Object.}

\item{sampleCol}{\code{character(1)}.
Column name of discrete samples to plot on X axis.}

\item{valueCol}{\code{character(1)}.
Column name of continues values to plot on Y axis.}

\item{interestingGroups}{\code{character}.
Groups of interest to use for visualization.
Corresponds to factors describing the columns of the object.}

\item{label}{\code{logical(1)}.
Include text labels showing the values above each bar.}

\item{fill}{\code{ggproto}/\code{ScaleDiscrete}.
Desired ggplot2 fill scale. Must supply discrete values. When set to
\code{NULL}, the default ggplot2 color palette will be used. If manual color
definitions are desired, we recommend using \code{\link[ggplot2:scale_manual]{ggplot2::scale_fill_manual()}}.

To set the discrete fill palette globally, use:\preformatted{options(acid.fill.discrete = ggplot2::scale_fill_viridis_d())
}}

\item{labels}{\code{list}.
ggplot2 labels.
See \code{\link[ggplot2:labs]{ggplot2::labs()}} for details.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Plot waterfall
}
\note{
Updated 2020-07-09.
}
\examples{
## data.frame ====
object <- data.frame(
    cell_id = paste("cell", seq_len(12L), sep = "_"),
    ic50 = seq(
        from = 0.1,
        to = 10L,
        length.out = 12L
    ),
    tumor_type = rep(
        x = c("breast", "bladder"),
        times = 6L
    ),
    tumor_subtype = rep(
        x = c("benign", "malignant"),
        each = 6L
    )
)
plotWaterfall(
    object = object,
    sampleCol = "cell_id",
    valueCol = "ic50",
    interestingGroups = c("tumor_type", "tumor_subtype"),
    labels = list(
        title = "Effect of compound on cell survival"
    )
)
plotWaterfall(
    object = object,
    sampleCol = "cell_id",
    valueCol = "ic50",
    label = FALSE
)
}
