% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotWaterfall-methods.R
\name{plotWaterfall}
\alias{plotWaterfall}
\alias{plotWaterfall,DataFrame-method}
\alias{plotWaterfall,SummarizedExperiment-method}
\alias{plotWaterfall,data.frame-method}
\title{Plot waterfall}
\usage{
plotWaterfall(object, ...)

\S4method{plotWaterfall}{DataFrame}(
  object,
  sampleCol,
  valueCol,
  interestingGroups = NULL,
  trans = c("identity", "log2", "log10"),
  fill = purpleOrange(1L),
  labels = list(title = NULL, subtitle = NULL)
)

\S4method{plotWaterfall}{SummarizedExperiment}(
  object,
  assay = 1L,
  fun = c("mean", "sum"),
  interestingGroups = NULL,
  ...
)

\S4method{plotWaterfall}{data.frame}(
  object,
  sampleCol,
  valueCol,
  interestingGroups = NULL,
  trans = c("identity", "log2", "log10"),
  fill = purpleOrange(1L),
  labels = list(title = NULL, subtitle = NULL)
)
}
\arguments{
\item{object}{Object.}

\item{sampleCol}{\code{character(1)}.
Column name of discrete samples to plot on X axis.}

\item{valueCol}{\code{character(1)}.
Column name of continues values to plot on Y axis.}

\item{interestingGroups}{\code{character}.
Groups of interest to use for visualization.
Corresponds to factors describing the columns of the object.}

\item{trans}{\code{character(1)}.
Name of the axis scale transformation to apply.

For more information:\preformatted{help(topic = "scale_x_continuous", package = "ggplot2")
}}

\item{fill}{\code{character(1)}.
R color name or hex color code (e.g. \code{"#AF52DE"}).}

\item{labels}{\code{list}.
ggplot2 labels.
See \code{\link[ggplot2:labs]{ggplot2::labs()}} for details.}

\item{assay}{\code{vector(1)}.
Assay name or index position.}

\item{fun}{\code{function}.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Plot waterfall
}
\note{
Updated 2021-09-10.
}
\examples{
data(RangedSummarizedExperiment, package = "AcidTest")

## data.frame ====
object <- data.frame(
    "cellId" = AcidGenerics::autopadZeros(
        object = paste("cell", seq_len(12L), sep = "_")
    ),
    "ic50" = seq(
        from = 0.1,
        to = 10L,
        length.out = 12L
    ),
    "tumorType" = rep(
        x = c("breast", "bladder"),
        times = 6L
    ),
    "tumorSubtype" = rep(
        x = c("benign", "malignant"),
        each = 6L
    )
)
plotWaterfall(
    object = object,
    sampleCol = "cellId",
    valueCol = "ic50",
    interestingGroups = c("tumorType", "tumorSubtype"),
    trans = "log10"
)

## SummarizedExperiment ====
object <- RangedSummarizedExperiment
plotWaterfall(object, trans = "identity")
}
\seealso{
\itemize{
\item \href{https://stackoverflow.com/questions/43176546}{Ordering bars from lowest to highest value in each facet}
\item \href{https://www.r-bloggers.com/2019/12/how-to-reorder-arrange-bars-with-in-each-facet-of-ggplot/}{Ordering bars within facets using tidytext}
}
}
