% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotZerosVsDepth-methods.R
\name{plotZerosVsDepth}
\alias{plotZerosVsDepth}
\alias{plotZerosVsDepth,SummarizedExperiment-method}
\title{Plot percentage of zeros vs. library depth}
\usage{
plotZerosVsDepth(object, ...)

\S4method{plotZerosVsDepth}{SummarizedExperiment}(
  object,
  assay = 1L,
  interestingGroups = NULL,
  labels = list(title = "Zeros vs. depth", subtitle = NULL, x = "library size (depth)",
    y = "dropout rate")
)
}
\arguments{
\item{object}{Object.}

\item{assay}{\code{vector(1)}.
Assay name or index position.}

\item{interestingGroups}{\code{character}.
Groups of interest to use for visualization.
Corresponds to factors describing the columns of the object.}

\item{labels}{\code{list}.
ggplot2 labels.
See \code{\link[ggplot2:labs]{ggplot2::labs()}} for details.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Plot percentage of zeros vs. library depth
}
\note{
Updated 2021-09-10.
}
\examples{
data(
    RangedSummarizedExperiment,
    SingleCellExperiment_splatter,
    package = "AcidTest"
)

## SummarizedExperiment ====
object <- RangedSummarizedExperiment
plotZerosVsDepth(object)

## SingleCellExperiment ====
object <- SingleCellExperiment_splatter
plotZerosVsDepth(object)
}
