% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotZerosVsDepth-methods.R
\docType{methods}
\name{plotZerosVsDepth}
\alias{plotZerosVsDepth}
\alias{plotZerosVsDepth,SummarizedExperiment-method}
\title{Plot percentage of zeros vs. library depth}
\usage{
plotZerosVsDepth(object, ...)

\S4method{plotZerosVsDepth}{SummarizedExperiment}(object, assay = 1L,
  interestingGroups = NULL, color = getOption(x =
  "acid.color.discrete", default =
  acidplots::scale_colour_synesthesia_d()), title = "Zeros vs. depth")
}
\arguments{
\item{object}{Object.}

\item{assay}{\code{vector(1)}.
Assay name or index position.}

\item{interestingGroups}{\code{character}.
Groups of interest to use for visualization.
Corresponds to factors describing the columns of the object.}

\item{color}{\code{ScaleDiscrete}.
Desired ggplot2 color scale. Must supply discrete values. When set \code{NULL},
the default ggplot2 color palette will be used. If manual color definitions
are desired, we recommend using \code{\link[ggplot2:scale_color_manual]{ggplot2::scale_color_manual()}}.

To set the discrete color palette globally, use:\preformatted{options(acid.color.discrete = ggplot2::scale_color_viridis_d())
}}

\item{title}{\code{character(1)}.
Title.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Plot percentage of zeros vs. library depth
}
\note{
Updated 2019-08-19.
}
\examples{
data(
    RangedSummarizedExperiment,
    SingleCellExperiment,
    package = "acidtest"
)
rse <- RangedSummarizedExperiment
sce <- SingleCellExperiment

## SummarizedExperiment ====
plotZerosVsDepth(rse)

## SingleCellExperiment ====
plotZerosVsDepth(sce)
}
