% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCounts-methods.R,
%   R/plotCountsPerBiotype-methods.R, R/plotCountsPerBroadClass-methods.R,
%   R/plotCountsPerGene-methods.R, R/plotGenderMarkers-methods.R,
%   R/plotGenesDetected-methods.R, R/plotPCA-methods.R, R/plotQC-methods.R,
%   R/plotTotalCounts-methods.R, R/plotZerosVsDepth-methods.R, R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{Gene2Symbol}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{basejump}{\code{\link[basejump]{Gene2Symbol}}}

  \item{BiocGenerics}{\code{\link[BiocGenerics]{plotPCA}}}

  \item{bioverbs}{\code{\link[bioverbs]{plotCounts}}, \code{\link[bioverbs]{plotCountsPerBiotype}}, \code{\link[bioverbs]{plotCountsPerBroadClass}}, \code{\link[bioverbs]{plotCountsPerGene}}, \code{\link[bioverbs]{plotGenderMarkers}}, \code{\link[bioverbs]{plotGenesDetected}}, \code{\link[bioverbs]{plotQC}}, \code{\link[bioverbs]{plotTotalCounts}}, \code{\link[bioverbs]{plotZerosVsDepth}}}
}}

