% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCorrelationHeatmap-methods.R,
%   R/plotHeatmap-methods.R, R/plotQuantileHeatmap-methods.R
\docType{import}
\name{reexports}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{bioverbs}{\code{\link[bioverbs]{plotCorrelationHeatmap}}, \code{\link[bioverbs]{plotHeatmap}}, \code{\link[bioverbs]{plotQuantileHeatmap}}}
}}

