% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{!!!}
\alias{!!}
\alias{aes}
\alias{after_stat}
\alias{ggplot}
\alias{ggsave}
\alias{inferno}
\alias{log_breaks}
\alias{magma}
\alias{plasma}
\alias{plot_annotation}
\alias{plot_layout}
\alias{percent}
\alias{pretty_breaks}
\alias{quo}
\alias{quo_text}
\alias{sym}
\alias{syms}
\alias{theme_set}
\alias{viridis}
\alias{wrap_plots}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ggplot2}{\code{\link[ggplot2]{aes}}, \code{\link[ggplot2:aes_eval]{after_stat}}, \code{\link[ggplot2]{ggplot}}, \code{\link[ggplot2]{ggsave}}, \code{\link[ggplot2:theme_get]{theme_set}}}

  \item{patchwork}{\code{\link[patchwork]{plot_annotation}}, \code{\link[patchwork]{plot_layout}}, \code{\link[patchwork]{wrap_plots}}}

  \item{rlang}{\code{\link[rlang:injection-operator]{!!}}, \code{\link[rlang:splice-operator]{!!!}}, \code{\link[rlang:defusing-advanced]{quo}}, \code{\link[rlang:quo_label]{quo_text}}, \code{\link[rlang]{sym}}, \code{\link[rlang:sym]{syms}}}

  \item{scales}{\code{\link[scales:breaks_log]{log_breaks}}, \code{\link[scales:label_percent]{percent}}, \code{\link[scales:breaks_pretty]{pretty_breaks}}}

  \item{viridis}{\code{\link[viridis:reexports]{inferno}}, \code{\link[viridis:reexports]{magma}}, \code{\link[viridis:reexports]{plasma}}, \code{\link[viridis:reexports]{viridis}}}
}}

