% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes-synesthesia.R
\name{synesthesia}
\alias{synesthesia}
\alias{synesthesia_pal}
\alias{scale_colour_synesthesia_c}
\alias{scale_color_synesthesia_c}
\alias{scale_colour_synesthesia_d}
\alias{scale_color_synesthesia_d}
\alias{scale_fill_synesthesia_c}
\alias{scale_fill_synesthesia_d}
\title{Synesthesia color palette}
\usage{
synesthesia(
  n = 256L,
  palette = c("lightPalette", "darkPalette", "draculaPalette", "macOSLightPalette",
    "macOSDarkPalette", "iOSLightPalette", "iOSDarkPalette")
)

synesthesia_pal(
  palette = c("lightPalette", "darkPalette", "draculaPalette", "macOSLightPalette",
    "macOSDarkPalette", "iOSLightPalette", "iOSDarkPalette")
)

scale_colour_synesthesia_c(
  ...,
  na.value = "grey50",
  guide = "colourbar",
  palette = c("lightPalette", "darkPalette", "draculaPalette", "macOSLightPalette",
    "macOSDarkPalette", "iOSLightPalette", "iOSDarkPalette")
)

scale_color_synesthesia_c(
  ...,
  na.value = "grey50",
  guide = "colourbar",
  palette = c("lightPalette", "darkPalette", "draculaPalette", "macOSLightPalette",
    "macOSDarkPalette", "iOSLightPalette", "iOSDarkPalette")
)

scale_colour_synesthesia_d(
  ...,
  palette = c("lightPalette", "darkPalette", "draculaPalette", "macOSLightPalette",
    "macOSDarkPalette", "iOSLightPalette", "iOSDarkPalette")
)

scale_color_synesthesia_d(
  ...,
  palette = c("lightPalette", "darkPalette", "draculaPalette", "macOSLightPalette",
    "macOSDarkPalette", "iOSLightPalette", "iOSDarkPalette")
)

scale_fill_synesthesia_c(
  ...,
  na.value = "grey50",
  guide = "colourbar",
  palette = c("lightPalette", "darkPalette", "draculaPalette", "macOSLightPalette",
    "macOSDarkPalette", "iOSLightPalette", "iOSDarkPalette")
)

scale_fill_synesthesia_d(
  ...,
  palette = c("lightPalette", "darkPalette", "draculaPalette", "macOSLightPalette",
    "macOSDarkPalette", "iOSLightPalette", "iOSDarkPalette")
)
}
\arguments{
\item{n}{\code{integer(1)}.
The number of colors (>= 1) to be in the palette.}

\item{palette}{\code{character(1)}.
Color palette name.}

\item{...}{Additional arguments.}

\item{na.value}{\code{character(1)}.
Missing values will be replaced with this value.}

\item{guide}{\code{character(1)} or \code{function}.
A function used to create a guide or its name.
See \code{ggplot2::guides} for details.}
}
\value{
\code{character} or \code{ggproto}.
}
\description{
Four color gradient defined by purple, blue, green, orange.
}
\details{
A little bit of an acid trip, especially when used for heatmaps.
}
\note{
Updated 2021-08-11.
}
\examples{
suppressPackageStartupMessages({
    library(ggplot2)
})

## Color, continuous.
cc <- ggplot(mpg, aes(cty, hwy)) +
    geom_jitter(aes(color = hwy))
cc + scale_color_synesthesia_c()

## Color, discrete.
cd <- ggplot(mpg, aes(fl)) +
    geom_bar(aes(color = fl), fill = NA)
cd + scale_color_synesthesia_d()

## Fill, continuous.
fc <- ggplot(diamonds, aes(carat, price)) +
    geom_hex()
fc + scale_fill_synesthesia_c()

## Fill, discrete.
fd <- ggplot(mpg, aes(fl)) +
    geom_bar(aes(fill = fl))
fd + scale_fill_synesthesia_d()

## Base color.
par(mar = rep(2.5, times = 4L))
image(
    matrix(data = seq_len(100L), nrow = 10, ncol = 10),
    col = synesthesia()
)
}
\seealso{
Adapted from viridis color palette:
\itemize{
\item \code{viridis::viridis()}.
\item \code{viridis::viridis_pal()}.
\item \code{viridis::scale_color_viridis_c()}.
\item \code{viridis::scale_color_viridis_d()}.
\item \code{viridis::scale_fill_viridis_c()}.
\item \code{viridis::scale_fill_viridis_d()}.
}

Useful color palette URLs:
\itemize{
\item http://www.stat.columbia.edu/~tzheng/files/Rcolor.pdf
\item https://matplotlib.org/tutorials/colors/colormaps.html
\item http://colorbrewer2.org/
\item http://colorspace.r-forge.r-project.org/
\item http://hclwizard.org/
\item https://developer.apple.com/design/human-interface-guidelines/
}

Additional color palette functions that may be relevant:
\itemize{
\item \code{grDevices::hsv()}, \code{grDevices::rgb()}, \code{grDevices::col2rgb()}.
\item \code{gplots::col2hex()}.
\item \code{ggplot2::continuous_scale()}, \code{ggplot2::discrete_scale()}.
\item \code{scales::gradient_n_pal()}.
\item \code{Seurat::CustomPalette()}, \code{Seurat::PurpleAndYellow()}.
\item \code{colorRamps::matlab.like()}, \code{colorRamps::matlab.like2()}.
\item \code{RColorBrewer::display.brewer.all()}.
\item \code{viridisLite::viridis.map}.
}
}
