% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-palettes.R
\name{synesthesia}
\alias{synesthesia}
\alias{synesthesia_pal}
\alias{scale_colour_synesthesia_c}
\alias{scale_colour_synesthesia_d}
\alias{scale_fill_synesthesia_c}
\alias{scale_fill_synesthesia_d}
\title{Synesthesia color palette}
\usage{
synesthesia(n = 256L)

synesthesia_pal()

scale_colour_synesthesia_c()

scale_colour_synesthesia_d()

scale_fill_synesthesia_c(..., na.value = "grey50", guide = "colourbar")

scale_fill_synesthesia_d(...)
}
\description{
Synesthesia color palette
}
\examples{
library(ggplot2)

p <- ggplot(diamonds, aes(carat, price)) + geom_hex()
p + scale_fill_viridis_c()
p + scale_fill_synesthesia_c()

image(matrix(1:400, 20), col = synesthesia())
}
\seealso{
Adapted from viridis color palette:
\itemize{
\item \code{viridis::viridis()}.
\item \code{viridis::viridis_pal()}.
\item \code{viridis::scale_colour_viridis_c()}.
\item \code{viridis::scale_colour_viridis_d()}.
\item \code{viridis::scale_fill_viridis_c()}.
\item \code{viridis::scale_fill_viridis_d()}.
}

Useful color palette URLs:
\itemize{
\item http://www.stat.columbia.edu/~tzheng/files/Rcolor.pdf
\item https://matplotlib.org/tutorials/colors/colormaps.html
\item https://www.nceas.ucsb.edu/~frazier/RSpatialGuides/colorPaletteCheatsheet.pdf
\item http://colorbrewer2.org/
\item http://colorspace.r-forge.r-project.org/
\item http://hclwizard.org/
}

Additional color palette functions that may be relevant:
\itemize{
\item \code{grDevices::hsv()}, \code{grDevices::rgb()}, \code{grDevices::col2rgb()}.
\item \code{gplots::col2hex()}.
\item \code{ggplot2::continuous_scale()}, \code{ggplot2::discrete_scale()}.
\item \code{scales::gradient_n_pal()}.
\item \code{colorRamps::matlab.like()}, \code{colorRamps::matlab.like2()}.
\item \code{viridisLite::viridis.map}.
\item \code{Seurat::CustomPalette()}, \code{Seurat::PurpleAndYellow()}.
\item \code{RColorBrewer::display.brewer.all()}.
}
}
