% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{themes}
\alias{themes}
\alias{acid_theme_light}
\alias{acid_theme_dark}
\title{ggplot2 themes}
\usage{
acid_theme_light(base_size = 14L, base_family = "", face = c("bold",
  "plain"), aspect_ratio = NULL, legend_position = c("right", "bottom",
  "top", "none"), grid = FALSE, minimal = FALSE)

acid_theme_dark(base_size = 14L, base_family = "", face = c("bold",
  "plain"), aspect_ratio = NULL, legend_position = c("right", "bottom",
  "top", "none"), grid = FALSE, minimal = FALSE)
}
\arguments{
\item{base_size}{\code{numeric(1)}. Base font size.}

\item{base_family}{\code{character(1)}. Base font family.}

\item{face}{\code{character(1)}. Font face ("\code{bold}", "\code{plain}").}

\item{aspect_ratio}{\code{numeric(1)}. Aspect ratio, specifying the plot
proportions. Use \code{1} for a perfectly square plot (including the axis
labels).}

\item{legend_position}{\code{character(1)}. Legend key position. We're being a
little more restrictive here, only allowing "\code{bottom}", "\code{right}", or
"\code{none}". Including the legend at the top or the left side of the plot
rarely makes sense and is discouraged.}

\item{grid}{\code{logical(1)}. Label the major panel grids with a gray accent.}

\item{minimal}{\code{logical(1)}. Remove all axis lines, axis ticks, and
panel borders.}
}
\value{
\code{theme}.
}
\description{
Complete \href{http://ggplot2.tidyverse.org}{ggplot2} themes.
}
\details{
Both themes are based off of \code{\link[ggplot2:theme_linedraw]{ggplot2::theme_linedraw()}}, but with
modifications and extra user-definable parameters.
}
\section{Light}{


High contrast black and white theme optimized for print. Recommended for
scientific manuscripts and website tutorials.
}

\section{Dark}{


Blackout theme that sets the plot background as black, with white text.
Inspired by \code{Seurat::DarkTheme()}, with some color modifications. Useful
for visualizing many points with a high dynamic color range, such as t-SNE
expression plots.
}

\examples{
library(ggplot2)
p <- ggplot(
    data = mpg,
    mapping = aes(
        x = manufacturer,
        y = displ,
        color = manufacturer,
        fill = manufacturer
    )
) +
    geom_point()

## Paperwhite theme.
p + acid_theme_light(legend_position = "none")

## Midnight theme.
p + acid_theme_dark(legend_position = "none")
}
\seealso{
\code{ggplot2::theme()}.
}
